Split test_agent_loop.py into focused modules (Phase 1, closes #426)
Extract 4,711 lines from tests/test_agent_loop.py into three new test
modules and a shared helper module, keeping the total pytest collection
count identical at 1078 tests:

- tests/agent_loop_helpers.py  — FakeRunner, builders, utility helpers
- tests/test_backends.py       — 37 Claude/Gemini/Codex backend tests
- tests/test_protocol.py       — 165 protocol parsing & validation tests
- tests/test_comment_rendering.py — 16 comment rendering tests

Update README.md to document the new test file layout and focused
subsystem run commands.
