+++ tests/test_agent_loop.py
@@ -181,6 +181,24 @@ from coding_review_agent_loop.workdir_guard import (
 from unittest.mock import MagicMock, patch
 
 
+
+from agent_loop_helpers import (
+    FakeRunner,
+    json_dumps,
+    command_index,
+    read_usage_summary,
+    prior_item_dispositions,
+    blocking_issues,
+    prior_plan_item_dispositions,
+    structured_pr_review,
+    structured_plan_review,
+    structured_plan_revision,
+    structured_plan_state,
+    structured_coder_followup,
+    make_config,
+    plan_decomposition_json,
+)
+
 @pytest.fixture(autouse=True)
 def _no_real_repair():
     """Prevent attempt_repair from calling the real Gemini CLI in all tests.
@@ -210,738 +228,6 @@ def _agent_commands_available(monkeypatch):
         return None
 
     monkeypatch.setattr(config_module.shutil, "which", which)
-
-
-class FakeRunner(Runner):
-    def __init__(
-        self,
-        *,
-        claude_outputs=None,
-        codex_outputs=None,
-        gemini_outputs=None,
-        antigravity_outputs=None,
-        issue_payload=None,
-        issue_comments=None,
-        pr_payload=None,
-        pr_check_runs_payload=None,
-        pr_status_payload=None,
-        pr_branch_protection_payload=None,
-        pr_branch_protection_returncode=0,
-        pr_branch_protection_stderr="",
-        pr_check_runs_returncode=0,
-        pr_check_runs_stderr="",
-        pr_status_returncode=0,
-        pr_status_stderr="",
-        repo_default_branch="main",
-        repo_default_branch_returncode=0,
-        git_status="",
-        git_remote="git@github.com:OWNER/REPO.git",
-        git_inside=True,
-        git_head="abc123",
-        tracked_files=None,
-        changed_files=None,
-        diff_returncode=0,
-        diff_stderr="",
-        issue_urls=None,
-        public_response_outputs=None,
-        advance_git_head_on_pr=True,
-    ):
-        super().__init__(dry_run=False)
-        self.claude_outputs = list(claude_outputs or [])
-        self.codex_outputs = list(codex_outputs or [])
-        self.gemini_outputs = list(gemini_outputs or [])
-        self.antigravity_outputs = list(antigravity_outputs or [])
-        self.issue_payload = {
-            "number": 56,
-            "state": "open",
-            "is_pr": False,
-            "url": "https://github.com/OWNER/REPO/issues/56",
-            "title": "Fix issue-mode context",
-            "body": "Original issue body.",
-        }
-        if issue_payload:
-            self.issue_payload.update(issue_payload)
-        self.issue_comments = list(issue_comments or [])
-        self.pr_payload = {
-            "number": 77,
-            "state": "OPEN",
-            "url": "https://github.com/OWNER/REPO/pull/77",
-            "title": "Improve review prompt context",
-            "body": "Fixes #56",
-            "headRefName": "feature/review-context",
-            "baseRefName": "main",
-            "headRefOid": "abc123",
-            "comments": [],
-            "reviews": [],
-        }
-        if pr_payload:
-            self.pr_payload.update(pr_payload)
-        self.pr_check_runs_payload = pr_check_runs_payload or {
-            "check_runs": [{"name": "test", "status": "completed", "conclusion": "success"}]
-        }
-        self.pr_status_payload = pr_status_payload or {"state": "success", "statuses": []}
-        self.pr_branch_protection_payload = pr_branch_protection_payload or {"contexts": ["test"]}
-        self.pr_branch_protection_returncode = pr_branch_protection_returncode
-        self.pr_branch_protection_stderr = pr_branch_protection_stderr
-        self.pr_check_runs_returncode = pr_check_runs_returncode
-        self.pr_check_runs_stderr = pr_check_runs_stderr
-        self.pr_status_returncode = pr_status_returncode
-        self.pr_status_stderr = pr_status_stderr
-        self.repo_default_branch = repo_default_branch
-        self.repo_default_branch_returncode = repo_default_branch_returncode
-        self.commands = []
-        self.comments = []
-        self.issues = []
-        self.git_status = git_status
-        self.git_remote = git_remote
-        self.git_inside = git_inside
-        self.git_head = git_head
-        self.tracked_files = tracked_files or [
-            "pyproject.toml",
-            "README.md",
-            "src/coding_review_agent_loop/cli.py",
-            "tests/test_agent_loop.py",
-        ]
-        self.changed_files = changed_files or ["src/coding_review_agent_loop/cli.py"]
-        self.diff_returncode = diff_returncode
-        self.diff_stderr = diff_stderr
-        self.issue_urls = list(issue_urls) if issue_urls is not None else None
-        self.public_response_outputs = list(public_response_outputs or [])
-        self.advance_git_head_on_pr = advance_git_head_on_pr
-        self._agent_pr_counter = 0
-
-    def _normalize_legacy_agent_output(self, output: str, prompt: str) -> str:
-        stripped = output.lstrip()
-        if stripped.startswith("{"):
-            try:
-                payload = json.loads(stripped)
-            except json.JSONDecodeError:
-                payload = None
-            if isinstance(payload, dict):
-                for key in ("response", "result"):
-                    value = payload.get(key)
-                    if isinstance(value, str):
-                        normalized_value = self._normalize_legacy_agent_output(value, prompt)
-                        if normalized_value != value:
-                            payload[key] = normalized_value
-                            return json.dumps(payload)
-            return output
-        signature_matches = re.findall(r"-- (OpenAI Codex|Google Gemini|Anthropic Claude)", prompt)
-        signature = signature_matches[-1] if signature_matches else "OpenAI Codex"
-        if (
-            '"kind": "pr_review"' in prompt
-            and '"kind": "coder_followup"' not in prompt
-            and "<!-- AGENT_STATE:" in output
-        ):
-            parsed = parse_review(output, reviewer="OpenAI Codex")
-            return structured_pr_review(
-                state=parsed.state,
-                summary=parsed.summary or "Review complete.",
-                blocking_items=[item.text for item in parsed.blocking_items],
-                same_pr_followups=[item.text for item in parsed.followups.same_pr],
-                future_followups=[item.text for item in parsed.followups.future],
-                prior_item_dispositions=[
-                    {
-                        key: value
-                        for key, value in {
-                            "item_id": item.item_id,
-                            "disposition": item.disposition,
-                            "note": item.note,
-                        }.items()
-                        if value is not None
-                    }
-                    for item in parsed.dispositions
-                ],
-                reviewer=signature,
-                human_requirements_resolved="<!-- HUMAN_REQUIREMENTS_RESOLVED -->" in output,
-            )
-        if (
-            '"kind": "plan_review"' in prompt
-            and '"kind": "plan_revision"' not in prompt
-            and "<!-- AGENT_PLAN_STATE:" in output
-        ):
-            state = parse_plan_state(output)
-            items = parse_plan_review_items(output, reviewer="OpenAI Codex")
-            future = [item.text for item in items.future] if state == "approved" else []
-            blocking = [item.text for item in items.blocking]
-            return structured_plan_review(
-                state=state,
-                summary=_review_freeform_summary_text(output) or "Plan review complete.",
-                blocking_plan_issues=blocking,
-                same_plan_followups=[item.text for item in items.same_plan],
-                future_followups=future,
-                prior_plan_item_dispositions=[
-                    {
-                        key: value
-                        for key, value in {
-                            "item_id": item.item_id,
-                            "disposition": item.disposition,
-                            "note": item.note,
-                        }.items()
-                        if value is not None
-                    }
-                    for item in parse_plan_item_dispositions(output, reviewer="OpenAI Codex")
-                ],
-                reviewer=signature,
-                human_requirements_resolved="<!-- HUMAN_REQUIREMENTS_RESOLVED -->" in output,
-            )
-        if '"kind": "plan_revision"' in prompt and "<!-- AGENT_PLAN_STATE:" in output:
-            return structured_plan_revision(
-                summary=_review_freeform_summary_text(output) or "Revised the plan.",
-                plan_steps=[
-                    line.strip("- ").strip()
-                    for line in output.splitlines()
-                    if line.strip().startswith("- ")
-                    and "Requirement " not in line
-                    and not line.strip().startswith("--")
-                ]
-                or [_review_freeform_summary_text(output) or "Revised the plan."],
-                human_requirements=(
-                    "\n" + HUMAN_REQUIREMENTS_ADDRESSED_MARKER
-                    if HUMAN_REQUIREMENTS_ADDRESSED_MARKER in output
-                    else ""
-                ),
-            )
-        if '"kind": "coder_followup"' in prompt and "<!-- AGENT_STATE:" in output:
-            item_ids = sorted(set(re.findall(r"\[(item-[A-Za-z0-9._-]+)\]", prompt)))
-            item_ids = [item_id for item_id in item_ids if item_id != HUMAN_REQUIREMENTS_ACK_ITEM_ID]
-            human_ids = sorted(set(re.findall(r"`(Requirement \d+)`|(?:^|\s)(Requirement \d+):", output)))
-            flattened_human_ids = [first or second for first, second in human_ids]
-            return structured_coder_followup(
-                state=parse_agent_state(output),
-                summary=_review_freeform_summary_text(output) or "Updated the PR.",
-                addressed_items=item_ids,
-                remaining_items=[],
-                human_requirement_ids=flattened_human_ids,
-            )
-        return output
-
-    def _next_agent_output(self, outputs):
-        output = outputs.pop(0)
-        if isinstance(output, dict):
-            return output
-        if isinstance(output, tuple):
-            return output
-        return output, 0
-
-    def _record_command(self, args, cwd):
-        cmd = [str(arg) for arg in args]
-        cwd_path = Path(cwd)
-        if not cwd_path.is_dir():
-            raise FileNotFoundError(cwd_path)
-        self.commands.append((cmd, cwd_path))
-        return cmd, cwd_path
-
-    def _maybe_write_public_response_file(self, cmd):
-        if not self.public_response_outputs:
-            return
-        prompt = "\n".join(cmd)
-        match = re.search(r"Write the final public response.*?\n\n([^\n]+/responses/[^\n]+\.md)", prompt, re.S)
-        if not match:
-            return
-        response_path = Path(match.group(1))
-        response_path.parent.mkdir(parents=True, exist_ok=True)
-        response = self.public_response_outputs.pop(0)
-        if isinstance(response, dict):
-            response = response.get("text", "")
-        elif isinstance(response, str):
-            response = self._normalize_legacy_agent_output(response, prompt)
-        response_path.write_text(response, encoding="utf-8")
-
-    def _maybe_advance_git_head_for_agent_pr(self, text: str) -> None:
-        if not self.advance_git_head_on_pr:
-            return
-        if "<!-- AGENT_PR:" not in text and "github.com/OWNER/REPO/pull/" not in text:
-            return
-        self._agent_pr_counter += 1
-        self.git_head = f"{self.git_head}-agent-{self._agent_pr_counter}"
-
-    def run_with_log(
-        self,
-        args,
-        *,
-        cwd,
-        log_path,
-        label,
-        progress_interval_seconds,
-        check=True,
-        env=None,
-        use_pty=False,
-        timeout_seconds=None,
-    ):
-        cmd, cwd_path = self._record_command(args, cwd)
-        log_path.parent.mkdir(parents=True, exist_ok=True)
-        ensure_log_dir_ignored(log_path.parent)
-
-        if cmd[:1] == ["claude"]:
-            output, returncode = self._next_agent_output(self.claude_outputs)
-            if isinstance(output, str):
-                output = self._normalize_legacy_agent_output(output, "\n".join(cmd))
-            self._maybe_write_public_response_file(cmd)
-            self._maybe_advance_git_head_for_agent_pr(output)
-            log_path.write_text(f"$ {' '.join(cmd)}\n\n{output}", encoding="utf-8")
-            return CommandResult(cmd, cwd_path, output, "", returncode)
-
-        if cmd[:2] == ["codex", "exec"]:
-            output = self._next_agent_output(self.codex_outputs)
-            if isinstance(output, dict):
-                public_response = output.get("public_response", "")
-                stdout = output.get("stdout", "")
-                returncode = output.get("returncode", 0)
-            else:
-                public_response, returncode = output
-                stdout = public_response
-            if isinstance(public_response, str):
-                normalized = self._normalize_legacy_agent_output(public_response, "\n".join(cmd))
-                if normalized != public_response:
-                    public_response = normalized
-            self._maybe_write_public_response_file(cmd)
-            if "--output-last-message" in cmd:
-                out_path = Path(cmd[cmd.index("--output-last-message") + 1])
-                out_path.write_text(public_response, encoding="utf-8")
-            self._maybe_advance_git_head_for_agent_pr(public_response)
-            log_path.write_text(f"$ {' '.join(cmd)}\n\ncodex completed", encoding="utf-8")
-            return CommandResult(cmd, cwd_path, stdout, "", returncode)
-
-        if cmd[:1] == ["gemini"]:
-            output = self._next_agent_output(self.gemini_outputs)
-            explicit_stdout = False
-            if isinstance(output, dict):
-                stdout = output.get("stdout", "")
-                returncode = output.get("returncode", 0)
-                explicit_stdout = True
-            else:
-                stdout, returncode = output
-            output = stdout
-            if isinstance(output, str) and not explicit_stdout:
-                output = self._normalize_legacy_agent_output(output, "\n".join(cmd))
-            self._maybe_write_public_response_file(cmd)
-            self._maybe_advance_git_head_for_agent_pr(output)
-            log_path.write_text(f"$ {' '.join(cmd)}\n\n{output}", encoding="utf-8")
-            return CommandResult(cmd, cwd_path, output, "", returncode)
-
-        if cmd[:1] == ["agy"]:
-            output = self._next_agent_output(self.antigravity_outputs)
-            if isinstance(output, dict):
-                stdout = output.get("stdout", "")
-                returncode = output.get("returncode", 0)
-            else:
-                stdout, returncode = output
-            self._maybe_write_public_response_file(cmd)
-            self._maybe_advance_git_head_for_agent_pr(stdout)
-            log_path.write_text(f"$ {' '.join(cmd)}\n\n{stdout}", encoding="utf-8")
-            return CommandResult(cmd, cwd_path, stdout, "", returncode)
-
-        return self.run(args, cwd=cwd, check=check)
-
-    def run(self, args, *, cwd, input_text=None, check=True, env=None):
-        cmd, cwd_path = self._record_command(args, cwd)
-
-        if cmd[:1] == ["claude"]:
-            output, returncode = self._next_agent_output(self.claude_outputs)
-            if isinstance(output, str):
-                output = self._normalize_legacy_agent_output(output, "\n".join(cmd))
-            self._maybe_advance_git_head_for_agent_pr(output)
-            return CommandResult(cmd, cwd_path, output, "", returncode)
-
-        if cmd[:2] == ["codex", "exec"]:
-            output = self._next_agent_output(self.codex_outputs)
-            if isinstance(output, dict):
-                public_response = output.get("public_response", "")
-                stdout = output.get("stdout", "")
-                returncode = output.get("returncode", 0)
-            else:
-                public_response, returncode = output
-                stdout = public_response
-            if isinstance(public_response, str):
-                normalized = self._normalize_legacy_agent_output(public_response, "\n".join(cmd))
-                if normalized != public_response:
-                    public_response = normalized
-            if "--output-last-message" in cmd:
-                out_path = Path(cmd[cmd.index("--output-last-message") + 1])
-                out_path.write_text(public_response, encoding="utf-8")
-            self._maybe_advance_git_head_for_agent_pr(public_response)
-            return CommandResult(cmd, cwd_path, stdout, "", returncode)
-
-        if cmd[:3] == ["gh", "pr", "comment"]:
-            if "--body-file" in cmd:
-                body_path = Path(cmd[cmd.index("--body-file") + 1])
-                raw_body = body_path.read_text(encoding="utf-8")
-            elif "--body" in cmd:
-                raw_body = cmd[cmd.index("--body") + 1]
-            else:
-                raw_body = ""
-            self.comments.append(_strip_round_metadata(raw_body))
-            self.pr_payload.setdefault("comments", []).append(
-                {
-                    "author": {"login": "coding-review-agent-loop"},
-                    "createdAt": f"2026-05-23T00:00:{len(self.pr_payload.get('comments', [])):02d}Z",
-                    "body": raw_body,
-                }
-            )
-            return CommandResult(cmd, cwd_path, "", "", 0)
-
-        if cmd[:3] == ["gh", "issue", "comment"]:
-            if "--body-file" in cmd:
-                body_path = Path(cmd[cmd.index("--body-file") + 1])
-                raw_body = body_path.read_text(encoding="utf-8")
-            elif "--body" in cmd:
-                raw_body = cmd[cmd.index("--body") + 1]
-            else:
-                raw_body = ""
-            self.comments.append(_strip_round_metadata(raw_body))
-            self.issue_comments.append(
-                {
-                    "author": {"login": "coding-review-agent-loop"},
-                    "createdAt": f"2026-05-23T00:00:{len(self.issue_comments):02d}Z",
-                    "body": raw_body,
-                }
-            )
-            return CommandResult(cmd, cwd_path, "", "", 0)
-
-        if cmd[:3] == ["gh", "issue", "create"]:
-            title = cmd[cmd.index("--title") + 1]
-            if "--body-file" in cmd:
-                body_path = Path(cmd[cmd.index("--body-file") + 1])
-                body = body_path.read_text(encoding="utf-8")
-            else:
-                body = cmd[cmd.index("--body") + 1]
-            self.issues.append({"title": title, "body": body})
-            if self.issue_urls is None:
-                issue_url = "https://github.com/OWNER/REPO/issues/99"
-            else:
-                issue_url = self.issue_urls.pop(0)
-            return CommandResult(cmd, cwd_path, f"{issue_url or ''}\n", "", 0)
-
-        if cmd[:3] == ["gh", "pr", "view"]:
-            if "--jq" in cmd and ".headRefOid" in cmd:
-                return CommandResult(cmd, cwd_path, "abc123\n", "", 0)
-            return CommandResult(cmd, cwd_path, json_dumps(self.pr_payload), "", 0)
-
-        if cmd[:3] == ["gh", "repo", "view"] and "defaultBranchRef" in cmd:
-            stdout = (
-                f"{self.repo_default_branch}\n"
-                if self.repo_default_branch_returncode == 0 and self.repo_default_branch
-                else ""
-            )
-            return CommandResult(
-                cmd,
-                cwd_path,
-                stdout,
-                "",
-                self.repo_default_branch_returncode,
-            )
-
-        if cmd[:3] == ["gh", "issue", "view"]:
-            payload = {
-                "number": self.issue_payload.get("number", 56),
-                "title": self.issue_payload.get("title"),
-                "body": self.issue_payload.get("body"),
-                "url": self.issue_payload.get("url"),
-                "author": self.issue_payload.get("author"),
-                "createdAt": self.issue_payload.get("createdAt"),
-                "comments": self.issue_comments,
-            }
-            return CommandResult(cmd, cwd_path, json_dumps(payload), "", 0)
-
-        if cmd[:2] == ["gh", "api"] and "/issues/" in cmd[2]:
-            return CommandResult(cmd, cwd_path, json_dumps(self.issue_payload), "", 0)
-
-        if cmd[:2] == ["gh", "api"] and cmd[2].endswith("/check-runs"):
-            if "--jq" in cmd:
-                return CommandResult(cmd, cwd_path, "success\n", "", 0)
-            stdout = (
-                json_dumps(self.pr_check_runs_payload) if self.pr_check_runs_returncode == 0 else ""
-            )
-            return CommandResult(
-                cmd,
-                cwd_path,
-                stdout,
-                self.pr_check_runs_stderr,
-                self.pr_check_runs_returncode,
-            )
-
-        if cmd[:2] == ["gh", "api"] and cmd[2].endswith("/status"):
-            stdout = json_dumps(self.pr_status_payload) if self.pr_status_returncode == 0 else ""
-            return CommandResult(
-                cmd,
-                cwd_path,
-                stdout,
-                self.pr_status_stderr,
-                self.pr_status_returncode,
-            )
-
-        if cmd[:2] == ["gh", "api"] and "/protection/required_status_checks" in cmd[2]:
-            stdout = (
-                json_dumps(self.pr_branch_protection_payload)
-                if self.pr_branch_protection_returncode == 0
-                else ""
-            )
-            return CommandResult(
-                cmd,
-                cwd_path,
-                stdout,
-                self.pr_branch_protection_stderr,
-                self.pr_branch_protection_returncode,
-            )
-
-        if cmd[:1] == ["sleep"]:
-            return CommandResult(cmd, cwd_path, "", "", 0)
-
-        if cmd[:3] == ["git", "rev-parse", "--is-inside-work-tree"]:
-            if self.git_inside:
-                return CommandResult(cmd, cwd_path, "true\n", "", 0)
-            return CommandResult(cmd, cwd_path, "false\n", "", 1)
-
-        if cmd[:3] == ["git", "rev-parse", "HEAD"]:
-            return CommandResult(cmd, cwd_path, f"{self.git_head}\n", "", 0)
-
-        if cmd[:3] == ["git", "checkout", "--detach"]:
-            if len(cmd) > 3 and cmd[3].startswith("refs/remotes/origin/pr/"):
-                self.git_head = self.pr_payload.get("headRefOid", self.git_head)
-            return CommandResult(cmd, cwd_path, "", "", 0)
-
-        if cmd[:2] == ["git", "ls-files"]:
-            return CommandResult(cmd, cwd_path, "\n".join(self.tracked_files) + "\n", "", 0)
-
-        if cmd[:3] == ["git", "diff", "--name-only"]:
-            stdout = "\n".join(self.changed_files) + "\n" if self.diff_returncode == 0 else ""
-            return CommandResult(cmd, cwd_path, stdout, self.diff_stderr, self.diff_returncode)
-
-        if cmd[:4] == ["git", "remote", "get-url", "origin"]:
-            return CommandResult(cmd, cwd_path, f"{self.git_remote}\n", "", 0)
-
-        if cmd[:3] == ["git", "status", "--porcelain"]:
-            return CommandResult(cmd, cwd_path, self.git_status, "", 0)
-
-        if cmd[:3] == ["gh", "repo", "clone"]:
-            Path(cmd[4]).mkdir(parents=True, exist_ok=True)
-            return CommandResult(cmd, cwd_path, "", "", 0)
-
-        return CommandResult(cmd, cwd_path, "", "", 0)
-
-
-def json_dumps(value):
-    import json
-
-    return json.dumps(value) + "\n"
-
-
-def command_index(commands, prefix, *, start=0):
-    for index in range(start, len(commands)):
-        cmd = commands[index][0]
-        if cmd[: len(prefix)] == prefix:
-            return index
-    raise AssertionError(f"Command with prefix {prefix!r} not found.")
-
-
-def read_usage_summary(log_dir: Path) -> dict:
-    summary_paths = list(log_dir.glob("*-usage-summary.json"))
-    assert len(summary_paths) == 1
-    return json.loads(summary_paths[0].read_text(encoding="utf-8"))
-
-
-def prior_item_dispositions(*lines: str) -> str:
-    if not lines:
-        return ""
-    return "\n\n### Prior unresolved item dispositions\n" + "\n".join(f"- {line}" for line in lines)
-
-
-def blocking_issues(*lines: str) -> str:
-    if not lines:
-        return ""
-    return "\n\n### Blocking issues\n" + "\n".join(f"- {line}" for line in lines)
-
-
-def prior_plan_item_dispositions(*lines: str) -> str:
-    if not lines:
-        return ""
-    return "\n\n### Prior unresolved plan item dispositions\n" + "\n".join(
-        f"- {line}" for line in lines
-    )
-
-
-def structured_pr_review(
-    *,
-    state: str = "approved",
-    summary: str = "Review complete.",
-    blocking_items: list[str] | None = None,
-    same_pr_followups: list[str] | None = None,
-    future_followups: list[str] | None = None,
-    prior_item_dispositions: list[dict[str, str]] | None = None,
-    reviewer: str = "OpenAI Codex",
-    human_requirements_resolved: bool = False,
-) -> str:
-    return (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": state,
-                "summary": summary,
-                "blocking_items": blocking_items or [],
-                "same_pr_followups": same_pr_followups or [],
-                "future_followups": future_followups or [],
-                "prior_item_dispositions": prior_item_dispositions or [],
-            }
-        )
-        + ("\n<!-- HUMAN_REQUIREMENTS_RESOLVED -->" if human_requirements_resolved else "")
-        + f"\n<!-- AGENT_STATE: {state} -->\n-- {reviewer}"
-    )
-
-
-def structured_plan_review(
-    *,
-    state: str = "approved",
-    summary: str = "Plan review complete.",
-    blocking_plan_issues: list[str] | None = None,
-    same_plan_followups: list[str] | None = None,
-    future_followups: list[str] | None = None,
-    prior_plan_item_dispositions: list[dict[str, str]] | None = None,
-    reviewer: str = "OpenAI Codex",
-    human_requirements_resolved: bool = False,
-) -> str:
-    return (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": state,
-                "summary": summary,
-                "blocking_plan_issues": blocking_plan_issues or [],
-                "same_plan_followups": same_plan_followups or [],
-                "future_followups": future_followups or [],
-                "prior_plan_item_dispositions": prior_plan_item_dispositions or [],
-            }
-        )
-        + ("\n<!-- HUMAN_REQUIREMENTS_RESOLVED -->" if human_requirements_resolved else "")
-        + f"\n<!-- AGENT_PLAN_STATE: {state} -->\n-- {reviewer}"
-    )
-
-
-def structured_plan_revision(
-    *,
-    summary: str = "Revised the plan.",
-    prior_plan_item_dispositions: list[dict[str, str]] | None = None,
-    plan_steps: list[str] | None = None,
-    reviewer: str = "Anthropic Claude",
-    human_requirements: str = "",
-) -> str:
-    return (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_revision",
-                "state": "blocking",
-                "summary": summary,
-                "prior_plan_item_dispositions": prior_plan_item_dispositions or [],
-                "plan_steps": plan_steps or ["Update the plan.", "Run the relevant tests."],
-            }
-        )
-        + human_requirements
-        + "\n<!-- AGENT_PLAN_STATE: blocking -->\n"
-        + f"-- {reviewer}"
-    )
-
-
-def structured_plan_state(
-    *,
-    state: str = "approved",
-    summary: str = "Implementation plan.",
-    plan_steps: list[str] | None = None,
-    reviewer: str = "Anthropic Claude",
-) -> str:
-    return (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_state",
-                "state": state,
-                "summary": summary,
-                "plan_steps": plan_steps or ["Update the code.", "Run the relevant tests."],
-            }
-        )
-        + f"\n<!-- AGENT_PLAN_STATE: {state} -->\n"
-        + f"-- {reviewer}"
-    )
-
-
-def structured_coder_followup(
-    *,
-    state: str = "blocking",
-    summary: str = "Updated the PR.",
-    addressed_items: list[str] | None = None,
-    remaining_items: list[str] | None = None,
-    addressed_item_notes: dict[str, str] | None = None,
-    remaining_item_notes: dict[str, str] | None = None,
-    human_requirement_ids: list[str] | None = None,
-    checked_discussion_directly: bool = False,
-    tests_run: list[str] | None = None,
-    reviewer: str = "Anthropic Claude",
-) -> str:
-    return (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": state,
-                "summary": summary,
-                "addressed_items": addressed_items or [],
-                "remaining_items": remaining_items or [],
-                "addressed_item_notes": addressed_item_notes or {},
-                "remaining_item_notes": remaining_item_notes or {},
-                "human_requirements": {
-                    "addressed_ids": human_requirement_ids or [],
-                    "checked_discussion_directly": checked_discussion_directly,
-                },
-                **({"tests_run": tests_run} if tests_run is not None else {}),
-            }
-        )
-        + f"\n<!-- AGENT_STATE: {state} -->\n-- {reviewer}"
-    )
-
-
-def make_config(tmp_path, *, create_dirs=True, **overrides):
-    config = {
-        "repo": "OWNER/REPO",
-        "claude_dir": tmp_path / "claude",
-        "codex_dir": tmp_path / "codex",
-        "gemini_dir": tmp_path / "gemini",
-        "coder": "claude",
-        "reviewer": "codex",
-        "base": "main",
-        "max_rounds": 5,
-        "auto_merge": False,
-        "dry_run": False,
-        "allow_shared_dir": False,
-        "claude_cmd": "claude",
-        "codex_cmd": "codex",
-        "gemini_cmd": "gemini",
-        "gh_cmd": "gh",
-        "claude_args": (),
-        "codex_args": (),
-        "gemini_args": (),
-        "test_command": None,
-        "pre_review_tests": True,
-        "ci_check_name": "test",
-        "ci_timeout_seconds": 1200,
-        "ci_poll_interval_seconds": 30,
-        "quiet": True,
-        "log_dir": tmp_path / "logs",
-        "progress_interval_seconds": 30,
-        "agent_max_retries": 2,
-        "agent_retry_backoff_seconds": (1, 1),
-        "agent_memory": True,
-        "refresh_agent_memory": False,
-        "agent_memory_dir": tmp_path / "claude" / ".agent-loop" / "memory",
-        "refresh_test_profile": False,
-    }
-    config.update(overrides)
-    if create_dirs:
-        config["claude_dir"].mkdir(parents=True, exist_ok=True)
-        config["codex_dir"].mkdir(parents=True, exist_ok=True)
-        config["gemini_dir"].mkdir(parents=True, exist_ok=True)
-    return AgentLoopConfig(**config)
 
 
 def test_workdir_guard_rejects_outside_home_path(tmp_path):
@@ -1089,28 +375,6 @@ def _make_migration(revision: str, down_revision: str | tuple[str, ...] | None) 
     )
 
 
-def plan_decomposition_json(*phases):
-    if not phases:
-        phases = (
-            {
-                "title": "Internal schema utilities",
-                "scope": "Add internal helpers only.",
-                "non_goals": "No live orchestrator behavior changes.",
-                "dependency_notes": "First phase; no dependencies.",
-                "rollout_risk": "low - internal only.",
-                "validation": "Run parser and orchestrator tests before the next phase.",
-                "parent_context": "Approved plan slice: add helpers and tests while preserving existing behavior.",
-                "automation": "agent-pr",
-                "depends_on": [],
-            },
-        )
-    return json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "plan_decomposition",
-            "phases": list(phases),
-        }
-    )
 
 
 def _init_git_checkout_with_origin(tmp_path: Path) -> Path:
@@ -1297,3135 +561,6 @@ def test_validate_pr_migration_topology_blocks_non_literal_changed_metadata(tmp_
     assert result.message is not None
     assert "Could not validate Alembic revision metadata" in result.message
     assert "e4f5a6b7c8d9_non_literal.py" in result.message
-
-
-def test_parse_claude_output_extracts_text_and_session_id():
-    raw = json.dumps({"result": "Hello.", "session_id": "abc123"})
-    text, sid, usage, raw_usage, model = _parse_claude_output(raw)
-    assert text == "Hello."
-    assert sid == "abc123"
-    assert usage is None
-    assert raw_usage is None
-    assert model is None
-
-
-def test_parse_claude_output_falls_back_on_plain_text():
-    raw = "plain response"
-    text, sid, usage, raw_usage, model = _parse_claude_output(raw)
-    assert text == "plain response"
-    assert sid is None
-    assert usage is None
-    assert raw_usage is None
-    assert model is None
-
-
-def test_parse_claude_output_falls_back_on_non_string_result():
-    raw = json.dumps({"result": 42, "session_id": "abc"})
-    text, sid, usage, raw_usage, model = _parse_claude_output(raw)
-    assert text == raw  # non-string result → fall back to raw
-    assert sid == "abc"
-    assert usage is None
-    assert raw_usage is None
-    assert model is None
-
-
-def test_parse_claude_output_extracts_model_from_model_usage():
-    raw = json.dumps({"result": "Hi.", "modelUsage": {"claude-sonnet-4-6": {"outputTokens": 5}}})
-    text, sid, usage, raw_usage, model = _parse_claude_output(raw)
-    assert text == "Hi."
-    assert model == "claude-sonnet-4-6"
-
-
-def test_parse_claude_output_extracts_primary_model_from_model_usage():
-    raw = json.dumps({
-        "result": "Reviewed.",
-        "modelUsage": {
-            "claude-haiku-4-5-20251001": {"inputTokens": 4750, "outputTokens": 20},
-            "claude-sonnet-4-6": {
-                "inputTokens": 18,
-                "outputTokens": 8263,
-                "cacheReadInputTokens": 715975,
-            },
-        },
-    })
-    text, sid, usage, raw_usage, model = _parse_claude_output(raw)
-    assert text == "Reviewed."
-    assert model == "claude-sonnet-4-6"
-
-
-def test_parse_claude_output_prefers_top_level_model_over_model_usage():
-    raw = json.dumps({
-        "result": "Reviewed.",
-        "model": "claude-opus-4-1",
-        "modelUsage": {
-            "claude-haiku-4-5-20251001": {"outputTokens": 20},
-            "claude-sonnet-4-6": {"outputTokens": 8263},
-        },
-    })
-    text, sid, usage, raw_usage, model = _parse_claude_output(raw)
-    assert text == "Reviewed."
-    assert model == "claude-opus-4-1"
-
-
-def test_parse_gemini_output_extracts_json_response():
-    raw = json.dumps({
-        "response": "Reviewed.\n<!-- AGENT_STATE: approved -->",
-        "session_id": "gemini-session-1",
-    })
-    text, sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text == "Reviewed.\n<!-- AGENT_STATE: approved -->"
-    assert sid == "gemini-session-1"
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_parse_gemini_output_falls_back_on_plain_text():
-    text, sid, usage, raw_usage, source = _parse_gemini_payload("plain response")
-    assert text == "plain response"
-    assert sid is None
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_parse_gemini_output_falls_back_on_non_string_response():
-    raw = json.dumps({"response": 42, "session_id": "gemini-session-1"})
-    text, sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text == raw
-    assert sid == "gemini-session-1"
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_parse_gemini_output_prefers_public_response_marker():
-    raw = f"""Warning: True color (24-bit) support not detected.
-YOLO mode is enabled. All tool calls will be automatically approved.
-I will inspect the PR before giving the final answer.
-Error executing tool read_file: Path not in workspace.
-{PUBLIC_RESPONSE_MARKER}
-## Review
-
-No blocking findings.
-
-<!-- AGENT_STATE: approved -->
-
--- Google Gemini
-"""
-    text, sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text.startswith("## Review")
-    assert "True color" not in text
-    assert "YOLO mode" not in text
-    assert "I will inspect" not in text
-    assert "Error executing tool" not in text
-    assert "<!-- AGENT_STATE: approved -->" in text
-    assert sid is None
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_parse_gemini_output_uses_last_public_response_marker():
-    raw = f"""Gemini may mention {PUBLIC_RESPONSE_MARKER} while planning.
-{PUBLIC_RESPONSE_MARKER}
-intermediate draft
-{PUBLIC_RESPONSE_MARKER}
-Final answer.
-<!-- AGENT_STATE: approved -->
-"""
-    text, _sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text == "Final answer.\n<!-- AGENT_STATE: approved -->\n"
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_parse_gemini_json_response_strips_public_response_marker():
-    raw = json.dumps({
-        "response": f"diagnostic\n{PUBLIC_RESPONSE_MARKER}\nReviewed.\n<!-- AGENT_STATE: approved -->",
-        "session_id": "gemini-session-1",
-    })
-    text, sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text == "Reviewed.\n<!-- AGENT_STATE: approved -->"
-    assert sid == "gemini-session-1"
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_parse_gemini_output_strips_cli_preamble_before_final_response():
-    raw = """Warning: True color (24-bit) support not detected.
-YOLO mode is enabled. All tool calls will be automatically approved.
-Attempt 1 failed with status 429. Retrying with backoff... _GaxiosError: [{
-  "error": {
-    "code": 429,
-    "message": "No capacity available for model gemini-3-flash-preview on the server"
-  }
-}]
-I am now ready to provide my final response.
-
----
-
-## Code Review
-
-Looks good.
-
-<!-- AGENT_STATE: approved -->
-
--- Google Gemini
-"""
-    text, sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text.startswith("## Code Review")
-    assert "_GaxiosError" not in text
-    assert "YOLO mode" not in text
-    assert "<!-- AGENT_STATE: approved -->" in text
-    assert sid is None
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_parse_gemini_output_strips_cli_preamble_before_plan_state_marker():
-    raw = """Warning: True color (24-bit) support not detected.
-YOLO mode is enabled.
-I will now review the plan.
-
----
-
-## Plan Review
-
-Looks like a solid approach.
-
-<!-- AGENT_PLAN_STATE: approved -->
-
--- Google Gemini
-"""
-    text, sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text.startswith("## Plan Review")
-    assert "YOLO mode" not in text
-    assert "<!-- AGENT_PLAN_STATE: approved -->" in text
-    assert sid is None
-
-
-def test_parse_gemini_output_preserves_markdown_rules_after_preamble():
-    raw = """Warning: True color (24-bit) support not detected.
-YOLO mode is enabled.
-
----
-
-## Summary
-
-Reviewed the change.
-
----
-
-## Details
-
-Still looks good.
-
-<!-- AGENT_STATE: approved -->
-"""
-    text, sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text.startswith("## Summary")
-    assert "YOLO mode" not in text
-    assert "## Details" in text
-    assert "\n---\n\n## Details" in text
-    assert sid is None
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_parse_gemini_output_strips_preamble_before_clarification_marker():
-    raw = """Warning: True color (24-bit) support not detected.
-I need to ask a question.
-
----
-
-    Which endpoint should I update?
-<!-- AGENT_CLARIFY -->
-"""
-    text, sid, usage, raw_usage, source = _parse_gemini_payload(raw)
-    assert text.startswith("    Which endpoint")
-    assert "True color" not in text
-    assert "<!-- AGENT_CLARIFY -->" in text
-    assert sid is None
-    assert usage is None
-    assert raw_usage is None
-
-
-def test_normalize_claude_usage_keeps_zero_cached_tokens_exact():
-    usage = _normalize_claude_usage(
-        {
-            "input_tokens": 12,
-            "cached_input_tokens": 0,
-            "output_tokens": 8,
-            "total_tokens": 20,
-        }
-    )
-
-    assert usage is not None
-    assert usage.mode == "exact"
-    assert usage.cached_input_tokens == 0
-
-
-def test_normalize_codex_usage_keeps_zero_reasoning_tokens():
-    usage = _normalize_codex_usage(
-        {
-            "input_tokens": 12,
-            "cached_input_tokens": 0,
-            "output_tokens": 8,
-            "reasoning_tokens": 0,
-            "total_tokens": 20,
-        }
-    )
-
-    assert usage is not None
-    assert usage.mode == "exact"
-    assert usage.reasoning_tokens == 0
-
-
-def test_normalize_gemini_usage_keeps_zero_token_values_exact():
-    usage = _normalize_gemini_usage(
-        {
-            "inputTokenCount": 0,
-            "cachedInputTokenCount": 0,
-            "outputTokenCount": 4,
-            "totalTokenCount": 4,
-        }
-    )
-
-    assert usage is not None
-    assert usage.mode == "exact"
-    assert usage.input_tokens == 0
-    assert usage.cached_input_tokens == 0
-
-
-def test_extract_codex_usage_reads_turn_completed_jsonl():
-    usage, raw_usage = _extract_codex_usage(
-        "\n".join(
-            [
-                json.dumps({"type": "turn.started"}),
-                json.dumps(
-                    {
-                        "type": "turn.completed",
-                        "usage": {
-                            "input_tokens": 120,
-                            "cached_input_tokens": 30,
-                            "output_tokens": 45,
-                            "reasoning_tokens": 11,
-                            "total_tokens": 206,
-                        },
-                    }
-                ),
-            ]
-        )
-    )
-
-    assert usage is not None
-    assert usage.mode == "exact"
-    assert usage.input_tokens == 120
-    assert usage.cached_input_tokens == 30
-    assert usage.output_tokens == 45
-    assert usage.reasoning_tokens == 11
-    assert usage.total_tokens == 206
-    assert raw_usage == {
-        "input_tokens": 120,
-        "cached_input_tokens": 30,
-        "output_tokens": 45,
-        "reasoning_tokens": 11,
-        "total_tokens": 206,
-    }
-
-
-def test_claude_backend_prefers_response_file_over_message_text(tmp_path):
-    runner = FakeRunner(
-        claude_outputs=[
-            json.dumps(
-                {
-                    "result": "stdout message text",
-                    "session_id": "claude-session-1",
-                }
-            )
-        ],
-        public_response_outputs=["response file text"],
-    )
-    config = make_config(tmp_path)
-
-    result = CLAUDE_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.response_file_text == "response file text"
-    assert result.message_text == "stdout message text"
-    assert result.text == "response file text"
-    assert result.session_id == "claude-session-1"
-
-
-def test_gemini_backend_prefers_response_file_over_message_text(tmp_path):
-    runner = FakeRunner(
-        gemini_outputs=[
-            json.dumps(
-                {
-                    "response": f"diagnostic\n{PUBLIC_RESPONSE_MARKER}\nstdout message text",
-                    "session_id": "gemini-session-1",
-                }
-            )
-        ],
-        public_response_outputs=["response file text"],
-    )
-    config = make_config(tmp_path)
-
-    result = GEMINI_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.response_file_text == "response file text"
-    assert result.message_text == "stdout message text"
-    assert result.text == "response file text"
-    assert result.session_id == "gemini-session-1"
-
-
-def test_codex_backend_prefers_response_file_over_last_message_and_stdout(tmp_path):
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "last message text",
-                "stdout": "\n".join(
-                    [
-                        "noisy stdout chatter",
-                        json.dumps(
-                            {
-                                "type": "turn.completed",
-                                "usage": {
-                                    "input_tokens": 12,
-                                    "cached_input_tokens": 3,
-                                    "output_tokens": 4,
-                                    "reasoning_tokens": 1,
-                                    "total_tokens": 20,
-                                },
-                            }
-                        ),
-                    ]
-                ),
-            }
-        ],
-        public_response_outputs=["response file text"],
-    )
-    config = make_config(tmp_path)
-
-    result = CODEX_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.response_file_text == "response file text"
-    assert result.message_text == "last message text"
-    assert result.text == "response file text"
-    assert result.usage is not None
-    assert result.usage.total_tokens == 20
-
-
-def test_codex_backend_prefers_last_message_over_stdout_without_response_file(tmp_path):
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "last message text",
-                "stdout": "raw stdout fallback",
-            }
-        ]
-    )
-    config = make_config(tmp_path)
-
-    result = CODEX_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.response_file_text is None
-    assert result.message_text == "last message text"
-    assert result.text == "last message text"
-
-
-def test_codex_backend_uses_stdout_when_files_are_absent_or_empty(tmp_path):
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "",
-                "stdout": "raw stdout fallback",
-            }
-        ]
-    )
-    config = make_config(tmp_path)
-
-    result = CODEX_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.response_file_text is None
-    assert result.message_text == "raw stdout fallback"
-    assert result.text == "raw stdout fallback"
-
-
-def test_codex_backend_dry_run_sets_message_text_without_response_file(tmp_path):
-    runner = FakeRunner(codex_outputs=[{"stdout": "dry run stdout"}])
-    config = make_config(tmp_path, dry_run=True)
-
-    result = CODEX_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.response_file_text is None
-    assert result.message_text == "dry run stdout"
-    assert result.text == "dry run stdout"
-
-
-def _write_codex_rollout(
-    codex_home: Path,
-    thread_id: str,
-    records: list[object],
-    *,
-    name_prefix: str = "rollout-2026-06-18T12-00-00",
-) -> Path:
-    rollout_path = (
-        codex_home
-        / "sessions"
-        / "2026"
-        / "06"
-        / "18"
-        / f"{name_prefix}-{thread_id}.jsonl"
-    )
-    rollout_path.parent.mkdir(parents=True, exist_ok=True)
-    rollout_path.write_text(
-        "\n".join(
-            record if isinstance(record, str) else json.dumps(record)
-            for record in records
-        ),
-        encoding="utf-8",
-    )
-    return rollout_path
-
-
-@pytest.mark.parametrize(
-    ("record", "expected"),
-    [
-        ({"payload": {"model": "gpt-5.5"}}, "gpt-5.5"),
-        (
-            {"turn": {"model": "gpt-5.5", "model_reasoning_effort": "medium"}},
-            "gpt-5.5 (medium)",
-        ),
-    ],
-)
-def test_codex_backend_detects_model_from_rollout(tmp_path, monkeypatch, record, expected):
-    thread_id = "019ed9d8-1111-7222-8333-444444444444"
-    codex_home = tmp_path / "codex-home"
-    _write_codex_rollout(codex_home, thread_id, [record])
-    monkeypatch.setenv("CODEX_HOME", str(codex_home))
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "last message text",
-                "stdout": json.dumps({"type": "thread.started", "thread_id": thread_id}),
-            }
-        ]
-    )
-    config = make_config(tmp_path)
-
-    result = CODEX_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.model_used == expected
-
-
-def test_codex_backend_parses_current_turn_context_rollout_schema(tmp_path, monkeypatch):
-    thread_id = "019ed9d8-1111-7222-8333-444444444444"
-    codex_home = tmp_path / "codex-home"
-    _write_codex_rollout(
-        codex_home,
-        thread_id,
-        [
-            {
-                "timestamp": "2026-06-18T12:00:00.000Z",
-                "type": "turn_context",
-                "payload": {
-                    "model": "gpt-5.5",
-                    "effort": "high",
-                },
-            }
-        ],
-    )
-    monkeypatch.setenv("CODEX_HOME", str(codex_home))
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "last message text",
-                "stdout": json.dumps({"type": "thread.started", "thread_id": thread_id}),
-            }
-        ]
-    )
-
-    result = CODEX_BACKEND.run(
-        runner,
-        make_config(tmp_path),
-        "Review this PR.",
-        run_id="run-1",
-    )
-
-    assert result.model_used == "gpt-5.5 (high)"
-
-
-def test_codex_backend_accepts_mixed_case_uuid_for_rollout_lookup(tmp_path, monkeypatch):
-    thread_id = "019ED9d8-1111-7222-8333-444444444444"
-    codex_home = tmp_path / "codex-home"
-    _write_codex_rollout(
-        codex_home,
-        thread_id,
-        [{"payload": {"model": "gpt-5.5", "effort": "medium"}}],
-    )
-    monkeypatch.setenv("CODEX_HOME", str(codex_home))
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "last message text",
-                "stdout": json.dumps({"type": "thread.started", "thread_id": thread_id}),
-            }
-        ]
-    )
-
-    result = CODEX_BACKEND.run(
-        runner,
-        make_config(tmp_path),
-        "Review this PR.",
-        run_id="run-1",
-    )
-
-    assert result.model_used == "gpt-5.5 (medium)"
-
-
-@pytest.mark.parametrize(
-    ("thread_id", "declared_model", "expected"),
-    [
-        ("*", None, None),
-        ("deadbeef-dead-beef-dead-beef", "gpt-5.4", "gpt-5.4"),
-    ],
-)
-def test_codex_backend_rejects_non_uuid_thread_id_before_rollout_lookup(
-    tmp_path,
-    monkeypatch,
-    thread_id,
-    declared_model,
-    expected,
-):
-    codex_home = tmp_path / "codex-home"
-    _write_codex_rollout(
-        codex_home,
-        thread_id,
-        [{"payload": {"model": "wrong-model"}}],
-    )
-    monkeypatch.setenv("CODEX_HOME", str(codex_home))
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "last message text",
-                "stdout": json.dumps({"type": "thread.started", "thread_id": thread_id}),
-            }
-        ]
-    )
-
-    result = CODEX_BACKEND.run(
-        runner,
-        make_config(tmp_path, codex_model=declared_model),
-        "Review this PR.",
-        run_id="run-1",
-    )
-
-    assert result.model_used == expected
-
-
-def test_codex_backend_declared_model_takes_precedence_over_rollout(tmp_path, monkeypatch):
-    thread_id = "019ed9d8-1111-7222-8333-444444444444"
-    codex_home = tmp_path / "codex-home"
-    _write_codex_rollout(
-        codex_home,
-        thread_id,
-        [{"payload": {"model": "gpt-5.5", "effort": "high"}}],
-    )
-    monkeypatch.setenv("CODEX_HOME", str(codex_home))
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "last message text",
-                "stdout": json.dumps({"type": "thread.started", "thread_id": thread_id}),
-            }
-        ]
-    )
-    config = make_config(
-        tmp_path,
-        codex_model="gpt-5.4",
-        codex_reasoning_effort="medium",
-    )
-
-    result = CODEX_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.model_used == "gpt-5.4 (medium)"
-
-
-@pytest.mark.parametrize(
-    ("stdout", "records"),
-    [
-        ("not json", [{"payload": {"model": "gpt-5.5"}}]),
-        (
-            json.dumps(
-                {
-                    "type": "thread.started",
-                    "thread_id": "019ed9d8-1111-7222-8333-444444444444",
-                }
-            ),
-            None,
-        ),
-        (
-            json.dumps(
-                {
-                    "type": "thread.started",
-                    "thread_id": "019ed9d8-1111-7222-8333-444444444444",
-                }
-            ),
-            ["not json", {"payload": {"model": ""}}, {"payload": {"model": 55}}],
-        ),
-    ],
-)
-def test_codex_backend_missing_or_invalid_rollout_model_falls_back_to_none(
-    tmp_path,
-    monkeypatch,
-    stdout,
-    records,
-):
-    thread_id = "019ed9d8-1111-7222-8333-444444444444"
-    codex_home = tmp_path / "codex-home"
-    if records is not None:
-        _write_codex_rollout(codex_home, thread_id, records)
-    monkeypatch.setenv("CODEX_HOME", str(codex_home))
-    runner = FakeRunner(
-        codex_outputs=[{"public_response": "last message text", "stdout": stdout}]
-    )
-    config = make_config(tmp_path)
-
-    result = CODEX_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.model_used is None
-
-
-def test_codex_backend_invalid_rollout_falls_back_to_declared_model(tmp_path, monkeypatch):
-    thread_id = "019ed9d8-1111-7222-8333-444444444444"
-    codex_home = tmp_path / "codex-home"
-    _write_codex_rollout(codex_home, thread_id, ["not json"])
-    monkeypatch.setenv("CODEX_HOME", str(codex_home))
-    runner = FakeRunner(
-        codex_outputs=[
-            {
-                "public_response": "last message text",
-                "stdout": json.dumps({"type": "thread.started", "thread_id": thread_id}),
-            }
-        ]
-    )
-    config = make_config(tmp_path, codex_model="gpt-5.4")
-
-    result = CODEX_BACKEND.run(runner, config, "Review this PR.", run_id="run-1")
-
-    assert result.model_used == "gpt-5.4"
-
-
-def test_parse_agent_state_accepts_html_marker():
-    assert parse_agent_state("looks fine\n<!-- AGENT_STATE: approved -->") == "approved"
-    assert parse_agent_state("needs work\n<!-- agent_state: BLOCKING -->") == "blocking"
-
-
-def test_parse_agent_state_uses_last_marker_as_authoritative():
-    text = """
-    Quoting earlier review: <!-- AGENT_STATE: blocking -->
-
-    Final decision:
-    <!-- AGENT_STATE: approved -->
-    """
-    assert parse_agent_state(text) == "approved"
-
-
-def test_parse_agent_state_requires_marker():
-    with pytest.raises(AgentLoopError):
-        parse_agent_state("LGTM")
-
-
-def test_parse_plan_state_uses_last_marker_as_authoritative():
-    text = """
-    Quoting earlier plan review: <!-- AGENT_PLAN_STATE: blocking -->
-
-    Final decision:
-    <!-- AGENT_PLAN_STATE: approved -->
-    """
-    assert parse_plan_state(text) == "approved"
-
-
-def test_parse_plan_state_requires_plan_marker():
-    with pytest.raises(AgentLoopError):
-        parse_plan_state("<!-- AGENT_STATE: approved -->")
-
-
-def test_parse_signed_human_requirement_body_extracts_text_before_signature():
-    body = parse_signed_human_requirement_body(
-        "Please use the absolute URL.\n\n-- Human Reviewer\n\nExtra text ignored."
-    )
-
-    assert body == "Please use the absolute URL."
-
-
-@pytest.mark.parametrize(
-    "signature",
-    [
-        "-- Human Reviewer",
-        "  -- Human Reviewer  ",
-        "-- human reviewer",
-        "-- HUMAN REVIEWER",
-    ],
-)
-def test_parse_signed_human_requirement_body_accepts_standalone_signature_variants(signature):
-    assert parse_signed_human_requirement_body(f"Required change.\n{signature}\n") == "Required change."
-
-
-@pytest.mark.parametrize(
-    "signature",
-    [
-        "-- OpenAI Codex",
-        "-- Anthropic Claude",
-        "-- Google Gemini",
-        "-- coding-review-agent-loop",
-        "Inline text -- Human Reviewer",
-    ],
-)
-def test_parse_signed_human_requirement_body_rejects_agent_and_non_standalone_signatures(
-    signature,
-):
-    assert parse_signed_human_requirement_body(f"Comment body.\n{signature}\n") is None
-
-
-def test_parse_non_blocking_followups_extracts_bullets_only_from_section():
-    review = """
-    Looks good.
-
-    ### Non-blocking follow-ups
-    - Add `.agent-loop/` to `.gitignore`.
-    1. Add regression coverage for stale memory refresh.
-       Include multiple reviewers.
-
-    ### Notes
-    - This is not a follow-up.
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    followups = parse_non_blocking_followups(review, reviewer="OpenAI Codex")
-
-    assert [(item.reviewer, item.text) for item in followups] == [
-        ("OpenAI Codex", "Add `.agent-loop/` to `.gitignore`."),
-        (
-            "OpenAI Codex",
-            "Add regression coverage for stale memory refresh. Include multiple reviewers.",
-        ),
-    ]
-
-
-def test_parse_approved_followups_extracts_same_pr_and_future_independently():
-    review = """
-    LGTM with cleanup.
-
-    ### Same-PR follow-ups
-    - Rename the helper for clarity.
-      Keep the public behavior unchanged.
-
-    ### Future follow-ups
-    1. Add an integration fixture later.
-
-    ### Non-blocking follow-ups
-    - Legacy future item.
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    followups = parse_approved_followups(review, reviewer="OpenAI Codex")
-
-    assert isinstance(followups.same_pr, tuple)
-    assert isinstance(followups.future, tuple)
-    assert [(item.reviewer, item.text) for item in followups.same_pr] == [
-        ("OpenAI Codex", "Rename the helper for clarity. Keep the public behavior unchanged.")
-    ]
-    assert [(item.reviewer, item.text) for item in followups.future] == [
-        ("OpenAI Codex", "Add an integration fixture later."),
-        ("OpenAI Codex", "Legacy future item."),
-    ]
-
-
-def test_parse_approved_followups_accepts_trailing_colons_on_headings():
-    review = """
-    LGTM with follow-ups.
-
-    ### Same-PR follow-ups:
-    - Rename the helper for clarity.
-
-    ### Future follow-ups:
-    - Add an integration fixture later.
-
-    ### Non-blocking follow-ups:
-    - Legacy future item.
-
-    <!-- AGENT_STATE: approved -->
-    -- Google Gemini
-    """
-
-    followups = parse_approved_followups(review, reviewer="Gemini")
-
-    assert [(item.reviewer, item.text) for item in followups.same_pr] == [
-        ("Gemini", "Rename the helper for clarity.")
-    ]
-    assert [(item.reviewer, item.text) for item in followups.future] == [
-        ("Gemini", "Add an integration fixture later."),
-        ("Gemini", "Legacy future item."),
-    ]
-
-
-@pytest.mark.parametrize(
-    ("same_pr_heading", "future_heading", "legacy_heading"),
-    [
-        (
-            "### **Same-PR follow-ups**",
-            "### **Future follow-ups**",
-            "### **Non-blocking follow-ups**",
-        ),
-        (
-            "### **Same-PR follow-ups**:",
-            "### **Future follow-ups.**",
-            "### **Non-blocking follow-ups:**",
-        ),
-        (
-            "### Same-PR follow-ups.",
-            "### Future follow-ups.",
-            "### Non-blocking follow-ups.",
-        ),
-    ],
-)
-def test_parse_approved_followups_accepts_common_markdown_heading_variants(
-    same_pr_heading, future_heading, legacy_heading
-):
-    review = f"""
-    LGTM with follow-ups.
-
-    {same_pr_heading}
-    - Rename the helper for clarity.
-
-    {future_heading}
-    - Add an integration fixture later.
-
-    {legacy_heading}
-    - Legacy future item.
-
-    <!-- AGENT_STATE: approved -->
-    -- Google Gemini
-    """
-
-    followups = parse_approved_followups(review, reviewer="Gemini")
-
-    assert [(item.reviewer, item.text) for item in followups.same_pr] == [
-        ("Gemini", "Rename the helper for clarity.")
-    ]
-    assert [(item.reviewer, item.text) for item in followups.future] == [
-        ("Gemini", "Add an integration fixture later."),
-        ("Gemini", "Legacy future item."),
-    ]
-
-
-def test_parse_approved_followups_stops_at_unrelated_bold_heading():
-    review = """
-    LGTM with follow-ups.
-
-    ### Future follow-ups
-    - Add an integration fixture later.
-
-    ### **Notes**
-    - This is not a follow-up.
-
-    <!-- AGENT_STATE: approved -->
-    -- Google Gemini
-    """
-
-    followups = parse_approved_followups(review, reviewer="Gemini")
-
-    assert [(item.reviewer, item.text) for item in followups.future] == [
-        ("Gemini", "Add an integration fixture later."),
-    ]
-
-
-def test_parse_approved_followups_extracts_bullets_and_prose_paragraphs():
-    bullet_review = """
-    Codex approves final pass.
-
-    ### Future follow-ups
-    - Refine token estimation for large review prompts.
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-    prose_review = """
-    Claude approves final pass.
-
-    ### Future follow-ups
-    The `_parse_gemini_output` helper is dead production code and could be removed
-    in a future cleanup.
-
-    ### Same-PR follow-ups
-    Rename the helper in this PR before merge.
-    Keep the behavior unchanged.
-
-    ### Notes
-    This note is outside the follow-up sections.
-
-    <!-- AGENT_STATE: approved -->
-    -- Anthropic Claude
-    """
-
-    bullet_followups = parse_approved_followups(bullet_review, reviewer="Codex")
-    prose_followups = parse_approved_followups(prose_review, reviewer="Claude")
-
-    assert [(item.reviewer, item.text) for item in bullet_followups.future] == [
-        ("Codex", "Refine token estimation for large review prompts."),
-    ]
-    assert [(item.reviewer, item.text) for item in prose_followups.future] == [
-        (
-            "Claude",
-            "The `_parse_gemini_output` helper is dead production code and could be removed in a future cleanup.",
-        ),
-    ]
-    assert [(item.reviewer, item.text) for item in prose_followups.same_pr] == [
-        ("Claude", "Rename the helper in this PR before merge. Keep the behavior unchanged."),
-    ]
-
-
-def test_parse_approved_followups_keeps_multiline_markdown_finding_as_one_item():
-    review = """
-    Still blocked.
-
-    ### Same-PR follow-ups
-    #### Normalize `_plan_subject` whitespace handling
-
-    Keep the helper from creating distinct round subjects for leading/trailing
-    whitespace-only differences.
-
-    ```python
-    assert _plan_subject("x") == _plan_subject(" x ")
-    ```
-
-    The implementation should preserve the current hash format.
-
-    ---
-
-    #### Harden `_decode_round_metadata` exception handling
-
-    Invalid base64 and invalid JSON should still become `AgentLoopError`
-    consistently.
-
-    ### Notes
-    CI note outside the section.
-
-    <!-- AGENT_STATE: blocking -->
-    -- Anthropic Claude
-    """
-
-    followups = parse_approved_followups(review, reviewer="Anthropic Claude")
-
-    assert [(item.reviewer, item.text) for item in followups.same_pr] == [
-        (
-            "Anthropic Claude",
-            "\n".join(
-                [
-                    "#### Normalize `_plan_subject` whitespace handling",
-                    "",
-                    "Keep the helper from creating distinct round subjects for leading/trailing whitespace-only differences.",
-                    "",
-                    "```python",
-                    'assert _plan_subject("x") == _plan_subject(" x ")',
-                    "```",
-                    "",
-                    "The implementation should preserve the current hash format.",
-                ]
-            ),
-        ),
-        (
-            "Anthropic Claude",
-            "\n".join(
-                [
-                    "#### Harden `_decode_round_metadata` exception handling",
-                    "",
-                    "Invalid base64 and invalid JSON should still become `AgentLoopError` consistently.",
-                ]
-            ),
-        ),
-    ]
-
-
-@pytest.mark.parametrize(
-    "placeholder",
-    [
-        "None",
-        "none.",
-        "(none)",
-        "(n/a)",
-        "N/A",
-        "No follow-ups",
-        "No same-PR follow-ups.",
-        "No future follow-ups",
-    ],
-)
-def test_parse_approved_followups_ignores_empty_placeholders(placeholder):
-    review = f"""
-    LGTM.
-
-    ### Same-PR follow-ups
-    - {placeholder}
-
-    ### Future follow-ups
-    - {placeholder}
-
-    <!-- AGENT_STATE: approved -->
-    -- Google Gemini
-    """
-
-    followups = parse_approved_followups(review, reviewer="Gemini")
-
-    assert followups.same_pr == ()
-    assert followups.future == ()
-
-
-def test_parse_approved_followups_ignores_prose_empty_placeholders():
-    review = """
-    LGTM.
-
-    ### Same-PR follow-ups
-    No same-PR follow-ups.
-
-    ### Future follow-ups
-    None
-
-    ### Notes
-    This sentence should not be captured.
-
-    <!-- AGENT_STATE: approved -->
-    -- Google Gemini
-    """
-
-    followups = parse_approved_followups(review, reviewer="Gemini")
-
-    assert followups.same_pr == ()
-    assert followups.future == ()
-
-
-def test_parse_plan_review_items_extracts_structured_sections():
-    review = """
-    Plan looks sound with one required revision.
-
-    ### Blocking plan issues
-    - Cover how the plan avoids mixing `AGENT_STATE` and `AGENT_PLAN_STATE`.
-
-    ### Same-plan follow-ups
-    - Mention the exact docs pages to update.
-
-    ### Future follow-ups
-    - Consider a later helper to unify plan and PR disposition rendering.
-
-    <!-- AGENT_PLAN_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    items = parse_plan_review_items(review, reviewer="OpenAI Codex")
-
-    assert [(item.reviewer, item.text) for item in items.blocking] == [
-        (
-            "OpenAI Codex",
-            "Cover how the plan avoids mixing `AGENT_STATE` and `AGENT_PLAN_STATE`.",
-        )
-    ]
-    assert [(item.reviewer, item.text) for item in items.same_plan] == [
-        ("OpenAI Codex", "Mention the exact docs pages to update.")
-    ]
-    assert [(item.reviewer, item.text) for item in items.future] == [
-        (
-            "OpenAI Codex",
-            "Consider a later helper to unify plan and PR disposition rendering.",
-        )
-    ]
-
-
-def test_parse_plan_review_items_keeps_multiline_markdown_blocking_item_as_one_entry():
-    review = """
-    Plan needs one revision.
-
-    ### Blocking plan issues
-    #### Preserve multiline review items during tracking
-
-    Do not split one reviewer-authored finding into separate ledger entries for
-    paragraphs or code blocks.
-
-    ```text
-    item-2: heading
-    item-3: paragraph
-    ```
-
-    ### Same-plan follow-ups
-    - Mention the regression shape in the implementation plan.
-
-    <!-- AGENT_PLAN_STATE: blocking -->
-    -- OpenAI Codex
-    """
-
-    items = parse_plan_review_items(review, reviewer="OpenAI Codex")
-
-    assert [(item.reviewer, item.text) for item in items.blocking] == [
-        (
-            "OpenAI Codex",
-            "\n".join(
-                [
-                    "#### Preserve multiline review items during tracking",
-                    "",
-                    "Do not split one reviewer-authored finding into separate ledger entries for paragraphs or code blocks.",
-                    "",
-                    "```text",
-                    "item-2: heading",
-                    "item-3: paragraph",
-                    "```",
-                ]
-            ),
-        )
-    ]
-
-
-@pytest.mark.parametrize(
-    "placeholder",
-    [
-        "None",
-        "(none)",
-        "(n/a)",
-        "No blocking plan issues.",
-        "No same-plan follow-ups",
-        "No future follow-ups.",
-    ],
-)
-def test_parse_plan_review_items_ignores_empty_placeholders(placeholder):
-    review = f"""
-    Looks good.
-
-    ### Blocking plan issues
-    - {placeholder}
-
-    ### Same-plan follow-ups
-    {placeholder}
-
-    ### Future follow-ups
-    - {placeholder}
-
-    <!-- AGENT_PLAN_STATE: approved -->
-    -- Google Gemini
-    """
-
-    items = parse_plan_review_items(review, reviewer="Gemini")
-
-    assert items.blocking == ()
-    assert items.same_plan == ()
-    assert items.future == ()
-
-
-def test_parse_plan_item_dispositions_extracts_same_plan_status():
-    review = """
-    Approved after the latest revision.
-
-    ### Prior unresolved plan item dispositions
-    - [item-1] resolved
-    - [item-2] still blocking
-    - [item-3] same-plan
-    - [item-4] future follow-up: okay to track separately now
-
-    <!-- AGENT_PLAN_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    dispositions = parse_plan_item_dispositions(review, reviewer="OpenAI Codex")
-
-    assert [(item.item_id, item.disposition, item.note) for item in dispositions] == [
-        ("item-1", "resolved", None),
-        ("item-2", "blocking", None),
-        ("item-3", "same-plan", None),
-        ("item-4", "future", "okay to track separately now"),
-    ]
-
-
-def test_parse_plan_item_dispositions_accepts_enriched_labels_with_trailing_arrow():
-    review = """
-    Approved after the latest revision.
-
-    ### Prior unresolved plan item dispositions
-    - [item-1] Same-plan follow-up from Google Gemini, round 1: keep the exact wording distinct -> same-plan: still need the mixed-reviewer case
-    - [item-2] Blocking issue from OpenAI Codex, round 1: preserve public labels -> resolved
-
-    <!-- AGENT_PLAN_STATE: blocking -->
-    -- Anthropic Claude
-    """
-
-    dispositions = parse_plan_item_dispositions(review, reviewer="Anthropic Claude")
-
-    assert [(item.item_id, item.disposition, item.note) for item in dispositions] == [
-        ("item-1", "same-plan", "still need the mixed-reviewer case"),
-        ("item-2", "resolved", None),
-    ]
-
-
-@pytest.mark.parametrize(
-    "line",
-    [
-        "[item-1] same-plan: none",
-        "[item-1] same-plan: N/A",
-        "[item-1] same-plan: no same-plan follow-ups",
-        "[item-1] still blocking: none",
-        "[item-1] still blocking: no blocking plan issues",
-        "[item-1] future follow-up: none",
-        "[item-1] future follow-up: no future follow-ups",
-    ],
-)
-def test_parse_plan_item_dispositions_rejects_contradictory_active_notes(line):
-    review = (
-        "Approved after the latest revision."
-        + prior_plan_item_dispositions(line)
-        + "\n\n<!-- AGENT_PLAN_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="use `resolved` when nothing remains"):
-        parse_plan_item_dispositions(review, reviewer="OpenAI Codex")
-
-
-@pytest.mark.parametrize("line", ["[item-1] same-plan:", "[item-1] still blocking:"])
-def test_parse_plan_item_dispositions_rejects_trailing_colon_syntax(line):
-    review = (
-        "Approved after the latest revision."
-        + prior_plan_item_dispositions(line)
-        + "\n\n<!-- AGENT_PLAN_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="Invalid prior unresolved plan item disposition"):
-        parse_plan_item_dispositions(review, reviewer="OpenAI Codex")
-
-
-def test_parse_plan_item_dispositions_allows_resolved_none_and_substantive_same_plan():
-    review = """
-    Approved after the latest revision.
-    """
-    review += prior_plan_item_dispositions(
-        "[item-1] resolved: none",
-        "[item-2] same-plan: still need the mixed-reviewer case",
-    )
-    review += "\n<!-- AGENT_PLAN_STATE: approved -->\n-- OpenAI Codex"
-
-    dispositions = parse_plan_item_dispositions(review, reviewer="OpenAI Codex")
-
-    assert [(item.item_id, item.disposition, item.note) for item in dispositions] == [
-        ("item-1", "resolved", "none"),
-        ("item-2", "same-plan", "still need the mixed-reviewer case"),
-    ]
-
-
-def test_parse_plan_item_dispositions_ignores_parenthesized_empty_placeholders():
-    review = """
-    Approved after the latest revision.
-
-    ### Prior unresolved plan item dispositions
-    - (none)
-    - (n/a)
-
-    <!-- AGENT_PLAN_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    assert parse_plan_item_dispositions(review, reviewer="OpenAI Codex") == ()
-
-
-def test_parse_plan_review_drops_future_followups_in_blocking_reviews():
-    review = structured_plan_review(
-        state="blocking",
-        summary="Still blocked.",
-        blocking_plan_issues=["Need clearer rollback coverage."],
-        future_followups=["Do this later."],
-    )
-
-    result = parse_plan_review(review, reviewer="OpenAI Codex")
-    assert result.items.future == ()
-    assert result.items.blocking  # blocking item survives
-
-
-def test_parse_plan_review_rejects_future_disposition_in_blocking_reviews():
-    review = structured_plan_review(
-        state="blocking",
-        summary="Still blocked.",
-        prior_plan_item_dispositions=[
-            {"item_id": "item-1", "disposition": "future", "note": "maybe later"}
-        ],
-    )
-
-    with pytest.raises(AgentLoopError, match="Blocking plan reviews may not downgrade"):
-        parse_plan_review(review, reviewer="OpenAI Codex")
-
-
-def test_parse_plan_review_rejects_contradictory_prior_plan_item_disposition():
-    review = structured_plan_review(
-        state="approved",
-        summary="Looks good now.",
-        prior_plan_item_dispositions=[
-            {"item_id": "item-1", "disposition": "same-plan", "note": "none"}
-        ],
-    )
-
-    with pytest.raises(AgentLoopError, match="empty placeholder"):
-        parse_plan_review(review, reviewer="OpenAI Codex")
-
-
-def test_parse_plan_review_rejects_approved_state_with_active_items():
-    plan_review = structured_plan_review(
-        state="approved",
-        summary="Needs work.",
-        same_plan_followups=["Add one more orchestration test."],
-    )
-
-    with pytest.raises(AgentLoopError, match="Approved plan reviews must be fully complete"):
-        parse_plan_review(plan_review, reviewer="OpenAI Codex")
-
-    with pytest.raises(AgentLoopError, match="AGENT_STATE"):
-        parse_review(plan_review, reviewer="OpenAI Codex")
-
-
-def test_parse_plan_review_rejects_approved_state_with_blocking_items():
-    review = structured_plan_review(
-        state="approved",
-        summary="Needs work.",
-        blocking_plan_issues=["Add one more orchestration test."],
-    )
-
-    with pytest.raises(AgentLoopError, match="Approved plan reviews must be fully complete"):
-        parse_plan_review(review, reviewer="OpenAI Codex")
-
-
-@pytest.mark.parametrize("line", ["[item-1] still blocking", "[item-1] same-plan"])
-def test_parse_plan_review_rejects_approved_state_with_active_prior_disposition(line):
-    item_id, disposition = ("item-1", "blocking") if "blocking" in line else ("item-1", "same-plan")
-    review = structured_plan_review(
-        state="approved",
-        summary="Looks good now.",
-        prior_plan_item_dispositions=[{"item_id": item_id, "disposition": disposition}],
-    )
-
-    with pytest.raises(AgentLoopError, match="Approved plan reviews must be fully complete"):
-        parse_plan_review(review, reviewer="OpenAI Codex")
-
-
-def test_validate_plan_review_response_rejects_duplicate_item_ids():
-    review = structured_plan_review(
-        state="blocking",
-        summary="Still refining the plan.",
-        prior_plan_item_dispositions=[
-            {"item_id": "item-1", "disposition": "same-plan", "note": "keep the extra regression coverage"},
-            {"item_id": "item-1", "disposition": "resolved"},
-        ],
-    )
-
-    with pytest.raises(AgentLoopError, match="more than once: item-1"):
-        _validate_plan_review_response(
-            review,
-            reviewer="OpenAI Codex",
-            unresolved_items=(
-                UnresolvedReviewItem(
-                    item_id="item-1",
-                    reviewer="Anthropic Claude",
-                    source_round=1,
-                    text="Keep the extra regression coverage.",
-                    status="same-plan",
-                ),
-            ),
-        )
-
-
-def test_validate_plan_review_response_rejects_unknown_item_ids():
-    review = structured_plan_review(
-        state="approved",
-        summary="Looks good now.",
-        prior_plan_item_dispositions=[{"item_id": "item-9", "disposition": "resolved"}],
-    )
-
-    with pytest.raises(UnknownPriorItemDispositionError, match="item-9"):
-        _validate_plan_review_response(
-            review,
-            reviewer="OpenAI Codex",
-            unresolved_items=(
-                UnresolvedReviewItem(
-                    item_id="item-1",
-                    reviewer="Anthropic Claude",
-                    source_round=1,
-                    text="Keep the extra regression coverage.",
-                    status="same-plan",
-                ),
-            ),
-        )
-
-
-def test_validate_plan_review_response_rejects_unknown_item_with_empty_prior_ledger():
-    review = structured_plan_review(
-        state="approved",
-        summary="Looks good now.",
-        prior_plan_item_dispositions=[{"item_id": "item-1", "disposition": "resolved"}],
-    )
-
-    with pytest.raises(UnknownPriorItemDispositionError) as exc_info:
-        _validate_plan_review_response(
-            review,
-            reviewer="OpenAI Codex",
-            unresolved_items=(),
-        )
-
-    assert exc_info.value.unknown_ids == ("item-1",)
-    assert exc_info.value.allowed_ids == ()
-    assert "Same-round findings are informational only" in exc_info.value.same_round_description
-
-
-def test_unknown_prior_item_disposition_error_message_includes_ids():
-    exc = UnknownPriorItemDispositionError(
-        unknown_ids=("item-15",),
-        allowed_ids=("item-12", "item-17"),
-        same_round_description="Same-round findings are informational only.",
-    )
-
-    message = str(exc)
-    assert "item-15" in message
-    assert "item-12" in message
-    assert "item-17" in message
-
-
-def test_validate_plan_review_response_describes_same_round_unknown_item():
-    review = structured_plan_review(
-        state="approved",
-        summary="Looks good now.",
-        prior_plan_item_dispositions=[{"item_id": "item-2", "disposition": "resolved"}],
-    )
-    current_round_item = UnresolvedReviewItem(
-        item_id="item-2",
-        reviewer="Google Gemini",
-        source_round=2,
-        text="Same-round finding.",
-        status="same-plan",
-    )
-
-    with pytest.raises(UnknownPriorItemDispositionError) as exc_info:
-        _validate_plan_review_response(
-            review,
-            reviewer="OpenAI Codex",
-            unresolved_items=(),
-            current_round_items=(current_round_item,),
-        )
-
-    assert "item-2" in exc_info.value.same_round_description
-
-
-def test_validate_plan_review_response_accepts_structured_resolved_dispositions():
-    review = structured_plan_review(
-        state="approved",
-        summary="Looks good now.",
-        prior_plan_item_dispositions=[
-            {"item_id": "item-1", "disposition": "resolved"},
-            {"item_id": "item-2", "disposition": "resolved"},
-        ],
-    )
-
-    parsed = _validate_plan_review_response(
-        review,
-        reviewer="OpenAI Codex",
-        unresolved_items=(
-            UnresolvedReviewItem(
-                item_id="item-1",
-                reviewer="Anthropic Claude",
-                source_round=1,
-                text="Keep the extra regression coverage.",
-                status="same-plan",
-            ),
-            UnresolvedReviewItem(
-                item_id="item-2",
-                reviewer="Google Gemini",
-                source_round=1,
-                text="Clarify the fallback trigger.",
-                status="blocking",
-            ),
-        ),
-    )
-
-    assert [(item.item_id, item.disposition) for item in parsed.dispositions] == [
-        ("item-1", "resolved"),
-        ("item-2", "resolved"),
-    ]
-
-
-def test_validate_plan_review_response_rejects_missing_structured_dispositions():
-    review = structured_plan_review(
-        state="approved",
-        summary="Looks good now.",
-        prior_plan_item_dispositions=[{"item_id": "item-1", "disposition": "resolved"}],
-    )
-
-    with pytest.raises(
-        AgentLoopError, match="did not evaluate all prior unresolved plan items: item-2"
-    ):
-        _validate_plan_review_response(
-            review,
-            reviewer="OpenAI Codex",
-            unresolved_items=(
-                UnresolvedReviewItem(
-                    item_id="item-1",
-                    reviewer="Anthropic Claude",
-                    source_round=1,
-                    text="Keep the extra regression coverage.",
-                    status="same-plan",
-                ),
-                UnresolvedReviewItem(
-                    item_id="item-2",
-                    reviewer="Google Gemini",
-                    source_round=1,
-                    text="Clarify the fallback trigger.",
-                    status="blocking",
-                ),
-            ),
-        )
-
-
-def test_parse_unresolved_item_dispositions_extracts_structured_updates():
-    review = """
-    LGTM.
-
-    ### Prior unresolved item dispositions
-    - [item-1] resolved
-    - [item-2] still blocking
-    - [item-3] same-pr
-    - [item-4] future follow-up: split this into a separate PR
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    dispositions = parse_unresolved_item_dispositions(review, reviewer="OpenAI Codex")
-
-    assert [(item.item_id, item.disposition, item.note) for item in dispositions] == [
-        ("item-1", "resolved", None),
-        ("item-2", "blocking", None),
-        ("item-3", "same-pr", None),
-        ("item-4", "future", "split this into a separate PR"),
-    ]
-
-
-def test_parse_unresolved_item_dispositions_accepts_enriched_labels_with_trailing_arrow():
-    review = """
-    LGTM.
-
-    ### Prior unresolved item dispositions
-    - [item-1] Same-PR follow-up from Google Gemini, round 1: require source issue reference in PR body -> same-pr: keep the body reference
-    - [item-2] Blocking issue from OpenAI Codex, round 1: rename the helper -> resolved
-
-    <!-- AGENT_STATE: blocking -->
-    -- Anthropic Claude
-    """
-
-    dispositions = parse_unresolved_item_dispositions(review, reviewer="Anthropic Claude")
-
-    assert [(item.item_id, item.disposition, item.note) for item in dispositions] == [
-        ("item-1", "same-pr", "keep the body reference"),
-        ("item-2", "resolved", None),
-    ]
-
-
-@pytest.mark.parametrize(
-    "line",
-    [
-        "[item-1] same-pr: none",
-        "[item-1] same-pr: N/A",
-        "[item-1] same-pr: no same-pr follow-ups",
-        "[item-1] still blocking: none",
-        "[item-1] still blocking: no blocking issues",
-        "[item-1] future follow-up: none",
-        "[item-1] future follow up: none",
-        "[item-1] future follow-up: no future follow-ups",
-        "[item-1] future follow-up: no follow-ups",
-    ],
-)
-def test_parse_unresolved_item_dispositions_rejects_contradictory_active_notes(line):
-    review = "LGTM." + prior_item_dispositions(line) + "\n\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-
-    with pytest.raises(AgentLoopError, match="use `resolved` when nothing remains"):
-        parse_unresolved_item_dispositions(review, reviewer="OpenAI Codex")
-
-
-@pytest.mark.parametrize("line", ["[item-1] same-pr:", "[item-1] still blocking:"])
-def test_parse_unresolved_item_dispositions_rejects_trailing_colon_syntax(line):
-    review = "LGTM." + prior_item_dispositions(line) + "\n\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-
-    with pytest.raises(
-        AgentLoopError,
-        match=r"Invalid prior unresolved item disposition.*section `### Prior unresolved item dispositions`, line 4",
-    ):
-        parse_unresolved_item_dispositions(review, reviewer="OpenAI Codex")
-
-
-def test_parse_unresolved_item_dispositions_allows_resolved_none_and_substantive_same_pr():
-    review = """
-    LGTM.
-    """
-    review += prior_item_dispositions(
-        "[item-1] resolved: none",
-        "[item-2] same-pr: rename the helper before merge",
-    )
-    review += "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-
-    dispositions = parse_unresolved_item_dispositions(review, reviewer="OpenAI Codex")
-
-    assert [(item.item_id, item.disposition, item.note) for item in dispositions] == [
-        ("item-1", "resolved", "none"),
-        ("item-2", "same-pr", "rename the helper before merge"),
-    ]
-
-
-def test_parse_unresolved_item_dispositions_ignores_parenthesized_empty_placeholders():
-    review = """
-    LGTM.
-
-    ### Prior unresolved item dispositions
-    - (none)
-    - (n/a)
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    assert parse_unresolved_item_dispositions(review, reviewer="OpenAI Codex") == ()
-
-
-def test_parse_unresolved_item_dispositions_ignores_non_bullet_prose():
-    review = """
-    LGTM.
-
-    ### Prior unresolved item dispositions
-    These are the remaining status calls.
-    - [item-1] resolved
-    Closing thought after the bullets.
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    dispositions = parse_unresolved_item_dispositions(review, reviewer="OpenAI Codex")
-
-    assert [(item.item_id, item.disposition, item.note) for item in dispositions] == [
-        ("item-1", "resolved", None),
-    ]
-
-
-def test_validate_review_response_accepts_structured_resolved_dispositions():
-    review = structured_pr_review(
-        state="approved",
-        summary="LGTM.",
-        prior_item_dispositions=[
-            {"item_id": "item-1", "disposition": "resolved"},
-            {"item_id": "item-2", "disposition": "resolved"},
-        ],
-    )
-
-    parsed = _validate_review_response(
-        review,
-        reviewer="OpenAI Codex",
-        unresolved_items=(
-            UnresolvedReviewItem(
-                item_id="item-1",
-                reviewer="Anthropic Claude",
-                source_round=1,
-                text="Rename the helper.",
-                status="same-pr",
-            ),
-            UnresolvedReviewItem(
-                item_id="item-2",
-                reviewer="Google Gemini",
-                source_round=1,
-                text="Keep the PR body issue reference.",
-                status="blocking",
-            ),
-        ),
-    )
-
-    assert [(item.item_id, item.disposition) for item in parsed.dispositions] == [
-        ("item-1", "resolved"),
-        ("item-2", "resolved"),
-    ]
-
-
-def test_validate_review_response_rejects_unknown_item_with_empty_prior_ledger():
-    review = structured_pr_review(
-        state="approved",
-        summary="LGTM.",
-        prior_item_dispositions=[{"item_id": "item-1", "disposition": "resolved"}],
-    )
-
-    with pytest.raises(UnknownPriorItemDispositionError) as exc_info:
-        _validate_review_response(
-            review,
-            reviewer="OpenAI Codex",
-            unresolved_items=(),
-        )
-
-    assert exc_info.value.unknown_ids == ("item-1",)
-    assert exc_info.value.allowed_ids == ()
-    assert "Same-round findings are informational only" in exc_info.value.same_round_description
-
-
-def test_validate_review_response_rejects_ambiguous_blanket_prose():
-    review = """
-    LGTM.
-
-    ### Prior unresolved item dispositions
-    All prior items look resolved.
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    with pytest.raises(AgentLoopError, match="required structured format"):
-        _validate_review_response(
-            review,
-            reviewer="OpenAI Codex",
-            unresolved_items=(
-                UnresolvedReviewItem(
-                    item_id="item-1",
-                    reviewer="Anthropic Claude",
-                    source_round=1,
-                    text="Rename the helper.",
-                    status="same-pr",
-                ),
-            ),
-        )
-
-
-def test_parse_review_drops_future_followups_in_blocking_reviews():
-    review = """
-    Still blocked.
-
-    ### Same-PR follow-ups
-    - Tighten the helper in this file.
-
-    ### Future follow-ups
-    - Do this later.
-
-    <!-- AGENT_STATE: blocking -->
-    -- OpenAI Codex
-    """
-
-    parsed = parse_review(review, reviewer="OpenAI Codex")
-
-    assert parsed.state == "blocking"
-    assert [item.text for item in parsed.followups.same_pr] == ["Tighten the helper in this file."]
-    assert parsed.followups.future == ()
-
-
-def test_parse_review_rejects_contradictory_prior_item_disposition():
-    review = "LGTM." + prior_item_dispositions("[item-1] same-pr: none")
-    review += "\n\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-
-    with pytest.raises(AgentLoopError, match="use `resolved` when nothing remains"):
-        parse_review(review, reviewer="OpenAI Codex")
-
-
-def test_parse_review_rejects_approved_state_with_same_pr_followups():
-    review = """
-    LGTM.
-
-    ### Same-PR follow-ups
-    - Tighten the helper in this file.
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    with pytest.raises(AgentLoopError, match="Approved reviews must be fully complete"):
-        parse_review(review, reviewer="OpenAI Codex")
-
-
-@pytest.mark.parametrize("line", ["[item-1] still blocking", "[item-1] same-pr"])
-def test_parse_review_rejects_approved_state_with_active_prior_disposition(line):
-    review = "LGTM." + prior_item_dispositions(line)
-    review += "\n\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-
-    with pytest.raises(AgentLoopError, match="Approved reviews must be fully complete"):
-        parse_review(review, reviewer="OpenAI Codex")
-
-
-def test_parse_review_populates_summary_from_legacy_markdown():
-    review = """
-    Blocking issue summary.
-
-    ### Same-PR follow-ups
-    - Rename the helper.
-
-    <!-- AGENT_STATE: blocking -->
-    -- OpenAI Codex
-    """
-
-    parsed = parse_review(review, reviewer="OpenAI Codex")
-
-    assert parsed.summary == "Blocking issue summary."
-
-
-def test_parse_review_round_trips_blocking_issues_section_without_polluting_summary():
-    review = (
-        "Blocking issue summary."
-        + blocking_issues(
-            "Cover the regression case in the PR test suite.",
-            "Tighten the error assertion wording.",
-        )
-        + "\n\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    parsed = parse_review(review, reviewer="OpenAI Codex")
-
-    assert parsed.summary == "Blocking issue summary."
-    assert [item.text for item in parsed.blocking_items] == [
-        "Cover the regression case in the PR test suite.",
-        "Tighten the error assertion wording.",
-    ]
-
-
-def test_parse_review_dedupes_same_pr_items_that_duplicate_blocking_items():
-    review = (
-        "Blocking issue summary."
-        + blocking_issues("`Add the missing share.html CSS update.`")
-        + "\n\n### Same-PR follow-ups\n"
-        + "- Add the missing share.html CSS update.\n"
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    parsed = parse_review(review, reviewer="OpenAI Codex")
-
-    assert [item.text for item in parsed.blocking_items] == [
-        "`Add the missing share.html CSS update.`"
-    ]
-    assert parsed.followups.same_pr == ()
-
-
-def test_parse_review_prefers_same_pr_over_duplicate_future_followups():
-    review = """
-    Blocking on a local cleanup.
-
-    ### Same-PR follow-ups
-    - Fix the duplicated prompt wording introduced by this PR.
-
-    ### Future follow-ups
-    - `Fix the duplicated prompt wording introduced by this PR.`
-
-    <!-- AGENT_STATE: blocking -->
-    -- OpenAI Codex
-    """
-
-    parsed = parse_review(review, reviewer="OpenAI Codex")
-
-    assert [item.text for item in parsed.followups.same_pr] == [
-        "Fix the duplicated prompt wording introduced by this PR."
-    ]
-    assert parsed.followups.future == ()
-
-
-def test_parse_review_prefers_blocking_over_duplicate_future_followups():
-    review = (
-        "Blocking issue summary."
-        + blocking_issues("Fix the indentation in the touched `orchestrator.py` call.")
-        + "\n\n### Future follow-ups\n"
-        + "- fix the indentation in the touched orchestrator.py call\n"
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    parsed = parse_review(review, reviewer="OpenAI Codex")
-
-    assert [item.text for item in parsed.blocking_items] == [
-        "Fix the indentation in the touched `orchestrator.py` call."
-    ]
-    assert parsed.followups.future == ()
-
-
-def test_parse_structured_pr_review_dedupes_exact_normalized_same_pr_duplicates():
-    review = json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "pr_review",
-            "state": "blocking",
-            "summary": "Blocked.",
-            "blocking_items": ["- Add the missing `share.html` CSS update."],
-            "same_pr_followups": ["Add the missing share.html CSS update"],
-            "future_followups": [],
-            "prior_item_dispositions": [],
-        }
-    )
-    review += "\n\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-
-    parsed = parse_pr_review(review, reviewer="OpenAI Codex")
-
-    assert [item.text for item in parsed.blocking_items] == [
-        "- Add the missing `share.html` CSS update."
-    ]
-    assert parsed.followups.same_pr == ()
-
-
-def test_parse_structured_pr_review_prefers_same_pr_over_duplicate_future_followups():
-    review = structured_pr_review(
-        state="blocking",
-        summary="Blocked on local cleanup.",
-        blocking_items=[],
-        same_pr_followups=["Fix the duplicated prompt wording introduced by this PR."],
-        future_followups=["fix the duplicated prompt wording introduced by this PR"],
-    )
-
-    parsed = parse_pr_review(review, reviewer="OpenAI Codex")
-
-    assert parsed.blocking_items == ()
-    assert [item.text for item in parsed.followups.same_pr] == [
-        "Fix the duplicated prompt wording introduced by this PR."
-    ]
-    assert parsed.followups.future == ()
-
-
-def test_parse_structured_pr_review_prefers_blocking_over_duplicate_future_followups():
-    review = structured_pr_review(
-        state="blocking",
-        summary="Blocked.",
-        blocking_items=["Fix the indentation in the touched `orchestrator.py` call."],
-        same_pr_followups=[],
-        future_followups=["fix the indentation in the touched orchestrator.py call"],
-    )
-
-    parsed = parse_pr_review(review, reviewer="OpenAI Codex")
-
-    assert [item.text for item in parsed.blocking_items] == [
-        "Fix the indentation in the touched `orchestrator.py` call."
-    ]
-    assert parsed.followups.same_pr == ()
-    assert parsed.followups.future == ()
-
-
-def test_parse_structured_pr_review_keeps_near_but_distinct_same_pr_items():
-    review = json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "pr_review",
-            "state": "blocking",
-            "summary": "Blocked.",
-            "blocking_items": ["Add the missing share.html CSS update."],
-            "same_pr_followups": ["Add the missing share.html print CSS update."],
-            "future_followups": [],
-            "prior_item_dispositions": [],
-        }
-    )
-    review += "\n\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-
-    parsed = parse_pr_review(review, reviewer="OpenAI Codex")
-
-    assert [item.text for item in parsed.blocking_items] == [
-        "Add the missing share.html CSS update."
-    ]
-    assert [item.text for item in parsed.followups.same_pr] == [
-        "Add the missing share.html print CSS update."
-    ]
-
-
-def test_pr_239_style_followup_classification_fixture():
-    review = """
-    PR #239-style cleanup classification.
-
-    ### Same-PR follow-ups
-    - orchestrator.py line 2100: subject=current_pr_subject is indented 4 extra spaces relative to sibling keyword arguments.
-    - _repair_prior_item_ids_instruction duplicates the same-round warning/context in the repair prompt.
-
-    ### Future follow-ups
-    - _round_ledger_may_be_incomplete cross-subject branch could be bounded if comment history grows very large.
-
-    <!-- AGENT_STATE: blocking -->
-    -- OpenAI Codex
-    """
-
-    followups = parse_approved_followups(review, reviewer="OpenAI Codex")
-
-    assert [item.text for item in followups.same_pr] == [
-        "orchestrator.py line 2100: subject=current_pr_subject is indented 4 extra spaces relative to sibling keyword arguments.",
-        "_repair_prior_item_ids_instruction duplicates the same-round warning/context in the repair prompt.",
-    ]
-    assert [item.text for item in followups.future] == [
-        "_round_ledger_may_be_incomplete cross-subject branch could be bounded if comment history grows very large."
-    ]
-
-
-def test_legacy_plan_review_helpers_populate_summary_from_markdown():
-    review = """
-    Plan needs one more regression test.
-
-    ### Same-plan follow-ups
-    - Add a regression test matrix.
-
-    <!-- AGENT_PLAN_STATE: blocking -->
-    -- OpenAI Codex
-    """
-
-    items = parse_plan_review_items(review, reviewer="OpenAI Codex")
-
-    assert _review_freeform_summary_text(review) == "Plan needs one more regression test."
-    assert [item.text for item in items.same_plan] == ["Add a regression test matrix."]
-
-
-def test_parse_plan_review_items_dedupes_plan_buckets_by_normalized_text():
-    review = """
-    Plan still needs cleanup.
-
-    ### Blocking plan issues
-    - Add `retry` coverage.
-
-    ### Same-plan follow-ups
-    - *add retry coverage!*
-    - Add parser comment.
-
-    ### Future follow-ups
-    - ADD RETRY COVERAGE.
-    - add `parser` comment
-    - Add parser documentation later.
-
-    <!-- AGENT_PLAN_STATE: blocking -->
-    -- OpenAI Codex
-    """
-
-    items = parse_plan_review_items(review, reviewer="OpenAI Codex")
-
-    assert [item.text for item in items.blocking] == ["Add `retry` coverage."]
-    assert [item.text for item in items.same_plan] == ["Add parser comment."]
-    assert [item.text for item in items.future] == ["Add parser documentation later."]
-
-
-def test_parse_structured_pr_review_normalizes_v1_payload_with_footer_contract():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "Looks good after the latest fix.",
-                "blocking_items": [],
-                "same_pr_followups": [],
-                "future_followups": ["Document cleanup for a later PR."],
-                "prior_item_dispositions": [
-                    {"item_id": "item-1", "disposition": "resolved"},
-                    {
-                        "item_id": "item-2",
-                        "disposition": "future",
-                        "note": "okay to split into follow-up work",
-                    },
-                ],
-            }
-        )
-        + "\n<!-- HUMAN_REQUIREMENTS_RESOLVED -->\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex\n"
-    )
-
-    parsed = parse_structured_pr_review(payload, reviewer="OpenAI Codex")
-
-    assert parsed is not None
-    assert parsed.state == "approved"
-    assert parsed.summary == "Looks good after the latest fix."
-    assert parsed.blocking_items == ()
-    assert [item.text for item in parsed.followups.future] == ["Document cleanup for a later PR."]
-    assert [(item.item_id, item.disposition, item.note) for item in parsed.dispositions] == [
-        ("item-1", "resolved", None),
-        ("item-2", "future", "okay to split into follow-up work"),
-    ]
-
-
-def test_parse_structured_pr_review_tolerates_omitted_empty_collections():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "Looks good after the latest fix.",
-                "prior_item_dispositions": [
-                    {"item_id": "item-1", "disposition": "resolved"},
-                ],
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex\n"
-    )
-
-    parsed = parse_structured_pr_review(payload, reviewer="OpenAI Codex")
-
-    assert parsed is not None
-    assert parsed.state == "approved"
-    assert parsed.blocking_items == ()
-    assert parsed.followups.same_pr == ()
-    assert parsed.followups.future == ()
-    assert [(item.item_id, item.disposition) for item in parsed.dispositions] == [
-        ("item-1", "resolved")
-    ]
-
-
-def test_parse_structured_pr_review_strips_verdict_and_sections_from_json_summary():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "blocking",
-                "summary": (
-                    "**Review verdict:** blocking\n\n"
-                    "Need one more regression test.\n\n"
-                    "### Blocking issues\n"
-                    "- Duplicate line that should not remain in the summary."
-                ),
-                "blocking_items": ["Need one more regression test."],
-                "same_pr_followups": [],
-                "future_followups": [],
-                "prior_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex\n"
-    )
-
-    parsed = parse_structured_pr_review(payload, reviewer="OpenAI Codex")
-
-    assert parsed is not None
-    assert parsed.summary == "Need one more regression test."
-
-
-def test_parse_structured_pr_review_rejects_kind_mismatch():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": "approved",
-                "summary": "Wrong kind.",
-                "blocking_plan_issues": [],
-                "same_plan_followups": [],
-                "future_followups": [],
-                "prior_plan_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="kind mismatch"):
-        parse_structured_pr_review(payload, reviewer="OpenAI Codex")
-
-
-def test_parse_structured_pr_review_hard_fails_on_unsupported_schema_version():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 2,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "Wrong version.",
-                "blocking_items": [],
-                "same_pr_followups": [],
-                "future_followups": [],
-                "prior_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="Unsupported structured response schema_version: 2"):
-        parse_structured_pr_review(payload, reviewer="OpenAI Codex")
-
-
-def test_parse_pr_review_rejects_markdown_when_no_structured_candidate_exists():
-    review = "Looks good in markdown.\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-
-    with pytest.raises(AgentLoopError, match="required structured format"):
-        parse_pr_review(review, reviewer="OpenAI Codex")
-
-
-def test_legacy_parse_review_still_parses_markdown_for_historical_display():
-    review = "Looks good in markdown.\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-
-    parsed = parse_review(review, reviewer="OpenAI Codex")
-
-    assert parsed.state == "approved"
-    assert parsed.summary == "Looks good in markdown."
-
-
-def test_parse_pr_review_rejects_invalid_structured_candidate_instead_of_falling_back_to_markdown():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "Missing required arrays.",
-                "blocking_items": [],
-                "same_pr_followups": [],
-                "future_followups": [],
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="missing required field"):
-        parse_pr_review(payload, reviewer="OpenAI Codex")
-
-
-def test_parse_structured_pr_review_strips_future_followups_in_blocking_reviews():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "blocking",
-                "summary": "Still blocked.",
-                "blocking_items": ["Needs one more test."],
-                "same_pr_followups": [],
-                "future_followups": ["Clean this up later."],
-                "prior_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    result = parse_structured_pr_review(payload, reviewer="OpenAI Codex")
-    assert result is not None
-    assert result.followups.future == ()
-    assert result.blocking_items  # blocking item survives
-
-
-def test_parse_pr_review_rejects_structured_candidate_with_unknown_nested_keys():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "LGTM.",
-                "blocking_items": [],
-                "same_pr_followups": [],
-                "future_followups": [],
-                "prior_item_dispositions": [
-                    {"item_id": "item-1", "disposition": "resolved", "extra": "nope"},
-                ],
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="unknown field"):
-        parse_pr_review(payload, reviewer="OpenAI Codex")
-
-
-def test_parse_pr_review_rejects_structured_candidate_with_invalid_item_id():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "LGTM.",
-                "blocking_items": [],
-                "same_pr_followups": [],
-                "future_followups": [],
-                "prior_item_dispositions": [
-                    {"item_id": "item 1", "disposition": "resolved"},
-                ],
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="must match"):
-        parse_pr_review(payload, reviewer="OpenAI Codex")
-
-
-def test_parse_pr_review_requires_strict_structured_disposition_enums():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "LGTM.",
-                "blocking_items": [],
-                "same_pr_followups": [],
-                "future_followups": [],
-                "prior_item_dispositions": [
-                    {"item_id": "item-1", "disposition": "still blocking"},
-                ],
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="must be one of"):
-        parse_pr_review(payload, reviewer="OpenAI Codex")
-
-
-def test_parse_structured_pr_review_rejects_approved_blocking_items():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "Almost there.",
-                "blocking_items": ["Still needs a regression test."],
-                "same_pr_followups": [],
-                "future_followups": [],
-                "prior_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="Approved reviews must be fully complete"):
-        parse_structured_pr_review(payload, reviewer="OpenAI Codex")
-
-
-@pytest.mark.parametrize(
-    "suffix",
-    [
-        "\nExtra explanation after the payload.\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex",
-        "\n```text\nextra block\n```\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex",
-        "\n- stray bullet\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex",
-    ],
-)
-def test_parse_structured_pr_review_rejects_trailing_content_before_footer(suffix):
-    payload = json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "pr_review",
-            "state": "approved",
-            "summary": "LGTM.",
-            "blocking_items": [],
-            "same_pr_followups": [],
-            "future_followups": [],
-            "prior_item_dispositions": [],
-        }
-    )
-
-    with pytest.raises(
-        AgentLoopError,
-        match="place <!-- AGENT_STATE|may not include prose between|may not include trailing prose",
-    ):
-        parse_structured_pr_review(payload + suffix, reviewer="OpenAI Codex")
-
-
-def test_parse_structured_pr_review_rejects_footer_state_mismatch():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "approved",
-                "summary": "LGTM.",
-                "blocking_items": [],
-                "same_pr_followups": [],
-                "future_followups": [],
-                "prior_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="must match the payload state"):
-        parse_structured_pr_review(payload, reviewer="OpenAI Codex")
-
-
-def test_parse_structured_pr_review_falls_back_when_json_is_embedded_in_markdown():
-    review = """
-    Here is an example:
-
-    ```json
-    {"schema_version": 1, "kind": "pr_review"}
-    ```
-
-    <!-- AGENT_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    assert parse_structured_pr_review(review, reviewer="OpenAI Codex") is None
-    assert parse_review(review, reviewer="OpenAI Codex").state == "approved"
-
-
-def test_parse_structured_plan_review_normalizes_v1_payload():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": "approved",
-                "summary": "Plan looks good.",
-                "blocking_plan_issues": [],
-                "same_plan_followups": [],
-                "future_followups": ["Consider a later cleanup pass."],
-                "prior_plan_item_dispositions": [{"item_id": "item-1", "disposition": "resolved"}],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    parsed = parse_structured_plan_review(payload, reviewer="OpenAI Codex")
-
-    assert parsed is not None
-    assert parsed.summary == "Plan looks good."
-    assert [item.text for item in parsed.items.future] == ["Consider a later cleanup pass."]
-    assert [(item.item_id, item.disposition) for item in parsed.dispositions] == [
-        ("item-1", "resolved")
-    ]
-
-
-def test_parse_structured_plan_review_tolerates_omitted_empty_collections():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": "approved",
-                "summary": "Plan looks good.",
-                "prior_plan_item_dispositions": [{"item_id": "item-1", "disposition": "resolved"}],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    parsed = parse_structured_plan_review(payload, reviewer="OpenAI Codex")
-
-    assert parsed is not None
-    assert parsed.items.blocking == ()
-    assert parsed.items.same_plan == ()
-    assert parsed.items.future == ()
-    assert [(item.item_id, item.disposition) for item in parsed.dispositions] == [
-        ("item-1", "resolved")
-    ]
-
-
-def test_parse_structured_plan_review_strips_verdict_and_sections_from_json_summary():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": "blocking",
-                "summary": (
-                    "**Review verdict:** blocking\n\n"
-                    "Need clearer rollback coverage.\n\n"
-                    "### Same-plan follow-ups\n"
-                    "- Extra duplicate text."
-                ),
-                "blocking_plan_issues": ["Need clearer rollback coverage."],
-                "same_plan_followups": [],
-                "future_followups": [],
-                "prior_plan_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    parsed = parse_structured_plan_review(payload, reviewer="OpenAI Codex")
-
-    assert parsed is not None
-    assert parsed.summary == "Need clearer rollback coverage."
-
-
-def test_parse_structured_plan_review_dedupes_same_plan_against_blocking_items():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": "blocking",
-                "summary": "Still blocked.",
-                "blocking_plan_issues": ["Add `retry` coverage."],
-                "same_plan_followups": [
-                    "*add retry coverage!*",
-                    "Add retry coverage for timeout handling.",
-                ],
-                "future_followups": [],
-                "prior_plan_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    parsed = parse_structured_plan_review(payload, reviewer="OpenAI Codex")
-
-    assert parsed is not None
-    assert [item.text for item in parsed.items.blocking] == ["Add `retry` coverage."]
-    assert [item.text for item in parsed.items.same_plan] == [
-        "Add retry coverage for timeout handling."
-    ]
-
-
-def test_parse_structured_plan_review_strips_blocking_future_followups():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": "blocking",
-                "summary": "Still blocked.",
-                "blocking_plan_issues": ["Need clearer rollback coverage."],
-                "same_plan_followups": [],
-                "future_followups": ["Refactor the prompt later."],
-                "prior_plan_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    result = parse_structured_plan_review(payload, reviewer="OpenAI Codex")
-    assert result is not None
-    assert result.items.future == ()
-    assert result.items.blocking  # blocking item survives
-
-
-def test_validate_structured_coder_followup_accepts_v1_payload():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Addressed the first item; one remains.",
-                "addressed_items": ["item-1"],
-                "remaining_items": ["item-2"],
-                "human_requirements": {
-                    "addressed_ids": ["Requirement 1"],
-                    "checked_discussion_directly": False,
-                },
-                "tests_run": ["python -m pytest tests/test_agent_loop.py -k structured"],
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    parsed = validate_structured_coder_followup(payload)
-
-    assert parsed is not None
-    assert parsed.addressed_items == ("item-1",)
-    assert parsed.remaining_items == ("item-2",)
-    assert parsed.human_requirements.addressed_ids == ("Requirement 1",)
-    assert parsed.addressed_item_notes == {}
-    assert parsed.remaining_item_notes == {}
-
-
-def test_validate_structured_coder_followup_accepts_optional_item_notes():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Addressed the parser; deferred the docs.",
-                "addressed_items": ["item-1"],
-                "remaining_items": ["item-2"],
-                "addressed_item_notes": {"item-1": "Added parsing coverage."},
-                "remaining_item_notes": {"item-2": "Deferred until the docs owner weighs in."},
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": False,
-                },
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    parsed = validate_structured_coder_followup(payload)
-
-    assert parsed is not None
-    assert parsed.addressed_item_notes == {"item-1": "Added parsing coverage."}
-    assert parsed.remaining_item_notes == {
-        "item-2": "Deferred until the docs owner weighs in."
-    }
-
-
-def test_validate_structured_coder_followup_rejects_note_for_unlisted_item():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Addressed one item.",
-                "addressed_items": ["item-1"],
-                "remaining_items": [],
-                "addressed_item_notes": {"item-2": "This note is stale."},
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": False,
-                },
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="item-2.*not listed in coder_followup.addressed_items"):
-        validate_structured_coder_followup(payload)
-
-
-@pytest.mark.parametrize("bad_note", ["", "   ", 5, None])
-def test_validate_structured_coder_followup_rejects_invalid_note_values(bad_note):
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Addressed one item.",
-                "addressed_items": ["item-1"],
-                "remaining_items": [],
-                "addressed_item_notes": {"item-1": bad_note},
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": False,
-                },
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="coder_followup.addressed_item_notes.item-1"):
-        validate_structured_coder_followup(payload)
-
-
-def test_validate_structured_coder_followup_returns_none_when_no_structured_candidate_exists():
-    assert (
-        validate_structured_coder_followup(
-            "Implemented the fix.\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-        )
-        is None
-    )
-
-
-def test_validate_structured_coder_followup_rejects_unknown_keys_in_structured_candidate():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "approved",
-                "summary": "Done.",
-                "addressed_items": [],
-                "remaining_items": [],
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": True,
-                    "extra": "nope",
-                },
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="unknown field\\(s\\): extra"):
-        validate_structured_coder_followup(payload)
-
-
-def test_validate_structured_coder_followup_rejects_footer_state_mismatch():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Done.",
-                "addressed_items": [],
-                "remaining_items": [],
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": True,
-                },
-            }
-        )
-        + "\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="footer AGENT_STATE must match the payload state"):
-        validate_structured_coder_followup(payload)
-
-
-def test_validate_structured_coder_followup_rejects_trailing_prose_after_footer():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Done.",
-                "addressed_items": [],
-                "remaining_items": [],
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": True,
-                },
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex\nextra"
-    )
-
-    with pytest.raises(AgentLoopError, match="may not include trailing prose"):
-        validate_structured_coder_followup(payload)
-
-
-@pytest.mark.parametrize(
-    ("addressed_ids", "checked_discussion_directly", "surfaced_ids", "requires_direct_discussion_ack", "message"),
-    [
-        (
-            ("Requirement 1",),
-            False,
-            ("Requirement 1", "Requirement 2"),
-            False,
-            "did not address all surfaced signed human requirement IDs",
-        ),
-        (
-            ("Requirement 1", "Requirement 1"),
-            False,
-            ("Requirement 1",),
-            False,
-            "listed signed human requirement IDs more than once",
-        ),
-        (
-            ("Requirement 99",),
-            False,
-            ("Requirement 1",),
-            False,
-            "referenced unknown signed human requirement IDs",
-        ),
-        (
-            (),
-            False,
-            (),
-            True,
-            "must acknowledge that the prompt omitted the detailed signed human requirements",
-        ),
-    ],
-)
-def test_validate_structured_human_requirements_acknowledgement_rejects_invalid_payloads(
-    addressed_ids,
-    checked_discussion_directly,
-    surfaced_ids,
-    requires_direct_discussion_ack,
-    message,
-):
-    with pytest.raises(AgentLoopError, match=message):
-        validate_structured_human_requirements_acknowledgement(
-            addressed_ids,
-            checked_discussion_directly=checked_discussion_directly,
-            surfaced_requirement_ids=surfaced_ids,
-            requires_direct_discussion_ack=requires_direct_discussion_ack,
-        )
-
-
-def test_validate_structured_plan_revision_accepts_v1_payload():
-    payload = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_revision",
-                "state": "blocking",
-                "summary": "Revised the plan to cover rollback testing.",
-                "prior_plan_item_dispositions": [
-                    {"item_id": "item-1", "disposition": "resolved", "note": "Covered in the new tests."}
-                ],
-                "plan_steps": ["Update protocol.py.", "Add regression tests."],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    parsed = validate_structured_plan_revision(payload)
-
-    assert parsed is not None
-    assert parsed.state == "blocking"
-    assert [(item.item_id, item.disposition) for item in parsed.prior_plan_item_dispositions] == [
-        ("item-1", "resolved")
-    ]
-    assert parsed.plan_steps == ("Update protocol.py.", "Add regression tests.")
-
-
-def test_validate_plan_revision_response_rejects_marker_only_markdown():
-    with pytest.raises(AgentLoopError, match="Plan revision did not use the required structured format"):
-        _validate_plan_revision_response(
-            "Revised plan.\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex"
-        )
-
-
-def test_validate_plan_revision_response_rejects_unknown_prior_disposition():
-    revision = structured_plan_revision(
-        prior_plan_item_dispositions=[
-            {"item_id": "item-15", "disposition": "resolved"},
-        ],
-    )
-    active_item = UnresolvedReviewItem(
-        item_id="item-12",
-        reviewer="Google Gemini",
-        source_round=5,
-        text="Active must-fix item.",
-        status="blocking",
-    )
-
-    with pytest.raises(UnknownPriorItemDispositionError) as exc_info:
-        _validate_plan_revision_response(revision, unresolved_items=(active_item,))
-
-    assert exc_info.value.unknown_ids == ("item-15",)
-    assert exc_info.value.allowed_ids == ("item-12",)
-
-
-def test_render_canonical_plan_revision_rejects_unknown_prior_disposition_without_keyerror():
-    revision = validate_structured_plan_revision(
-        structured_plan_revision(
-            prior_plan_item_dispositions=[
-                {"item_id": "item-15", "disposition": "resolved"},
-            ],
-        )
-    )
-    assert revision is not None
-    active_item = UnresolvedReviewItem(
-        item_id="item-12",
-        reviewer="Google Gemini",
-        source_round=5,
-        text="Active must-fix item.",
-        status="blocking",
-    )
-
-    with pytest.raises(AgentLoopError, match="Renderer encountered unknown prior item ID") as exc_info:
-        render_canonical_plan_revision(revision, prior_items=(active_item,))
-
-    assert not isinstance(exc_info.value, KeyError)
-
-
-@pytest.mark.parametrize(
-    ("payload", "pattern"),
-    [
-        (
-            {
-                "schema_version": 1,
-                "kind": "plan_revision",
-                "state": "approved",
-                "summary": "Wrong state.",
-                "prior_plan_item_dispositions": [],
-                "plan_steps": ["Update protocol.py."],
-            },
-            "plan_revision.state must be `blocking`",
-        ),
-        (
-            {
-                "schema_version": 1,
-                "kind": "plan_revision",
-                "state": "blocking",
-                "summary": "",
-                "prior_plan_item_dispositions": [],
-                "plan_steps": ["Update protocol.py."],
-            },
-            "plan_revision.summary must be a non-empty string",
-        ),
-        (
-            {
-                "schema_version": 1,
-                "kind": "plan_revision",
-                "state": "blocking",
-                "summary": "Missing steps.",
-                "prior_plan_item_dispositions": [],
-                "plan_steps": [],
-            },
-            "plan_revision.plan_steps must contain at least 1 item",
-        ),
-    ],
-)
-def test_validate_structured_plan_revision_rejects_invalid_payload(payload, pattern):
-    footer_state = payload["state"]
-    text = json.dumps(payload) + f"\n<!-- AGENT_PLAN_STATE: {footer_state} -->\n-- OpenAI Codex"
-
-    with pytest.raises(AgentLoopError, match=pattern):
-        validate_structured_plan_revision(text)
-
-
-def test_extract_structured_plan_review_payload_rejects_embedded_json_markdown():
-    review = """
-    Here is an example:
-
-    ```json
-    {"schema_version": 1, "kind": "plan_review"}
-    ```
-
-    <!-- AGENT_PLAN_STATE: approved -->
-    -- OpenAI Codex
-    """
-
-    assert _extract_structured_plan_review_payload(review) is None
-
-
-@pytest.mark.parametrize(
-    ("builder", "extractor"),
-    [
-        (lambda: structured_plan_review(reviewer="Google Gemini"), _extract_structured_plan_review_payload),
-        (lambda: structured_pr_review(reviewer="Google Gemini"), _extract_structured_pr_review_payload),
-        (lambda: structured_coder_followup(reviewer="Anthropic Claude"), _extract_structured_coder_followup_payload),
-        (lambda: structured_plan_revision(reviewer="Anthropic Claude"), _extract_structured_plan_revision_payload),
-    ],
-)
-def test_structured_extractors_recover_leading_public_response_marker(builder, extractor):
-    text = f"\n\n{PUBLIC_RESPONSE_MARKER}\n{builder()}"
-
-    payload = extractor(text)
-
-    assert payload is not None
-
-
-def test_response_file_marker_normalization_reports_unrecoverable_marker():
-    text = f"{PUBLIC_RESPONSE_MARKER}\n### Review\nLooks good."
-
-    normalized, status = normalize_response_file_structured_text(text)
-
-    assert normalized == text
-    assert status == "leading-public-response-marker-not-recoverable"
-
-
-def test_public_response_json_prefix_strips_marker_variants():
-    text = "==== AGENT_LOOP_PUBLIC_RESPONSE_BELOW ====\n" + json.dumps(
-        {"error": {"status": 429, "message": "quota exceeded"}}
-    )
-
-    payload = _decode_public_response_json_prefix(text)
-
-    assert payload == {"error": {"status": 429, "message": "quota exceeded"}}
-    assert _is_transient_public_response(text)
-
-
-def test_failure_category_threads_public_response_expected_kind():
-    text = json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "plan_review",
-            "message": "quota exceeded",
-        }
-    )
-
-    assert _failure_category(text, public_response=True) == "deterministic"
-    assert (
-        _failure_category(
-            text,
-            public_response=True,
-            repair_expected_kind="future_structured_kind",
-        )
-        == "transient"
-    )
-
-
-@pytest.mark.parametrize(
-    "text",
-    [
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": "approved",
-                "summary": PUBLIC_RESPONSE_MARKER,
-                "blocking_plan_issues": [],
-                "same_plan_followups": [],
-                "future_followups": [],
-                "prior_plan_item_dispositions": [],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: approved -->\n-- Google Gemini",
-        "Some prose first.\n"
-        + PUBLIC_RESPONSE_MARKER
-        + "\n"
-        + structured_plan_review(reviewer="Google Gemini"),
-    ],
-)
-def test_response_file_marker_not_stripped_inside_json_or_mid_prose(text):
-    normalized, status = normalize_response_file_structured_text(text)
-
-    assert normalized == text
-    assert status is None
-
-
-def test_extract_structured_plan_review_payload_rejects_footer_state_mismatch():
-    payload = json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "plan_review",
-            "state": "approved",
-            "summary": "Plan looks good.",
-            "blocking_plan_issues": [],
-            "same_plan_followups": [],
-            "future_followups": [],
-            "prior_plan_item_dispositions": [],
-        }
-    )
-    text = payload + "\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex"
-
-    with pytest.raises(AgentLoopError, match="footer AGENT_PLAN_STATE must match"):
-        _extract_structured_plan_review_payload(text)
-
-
-def test_extract_structured_plan_review_payload_rejects_trailing_prose_after_signature():
-    payload = json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "plan_review",
-            "state": "approved",
-            "summary": "Plan looks good.",
-            "blocking_plan_issues": [],
-            "same_plan_followups": [],
-            "future_followups": [],
-            "prior_plan_item_dispositions": [],
-        }
-    )
-    text = payload + "\n<!-- AGENT_PLAN_STATE: approved -->\n-- OpenAI Codex\nextra"
-
-    with pytest.raises(AgentLoopError, match="trailing prose"):
-        _extract_structured_plan_review_payload(text)
-
-
-def test_parse_plan_review_hard_fails_after_top_level_json_prefix():
-    review = (
-        '{"schema_version":1,"kind":"plan_review","state":"approved","summary":"Plan looks good.",'
-        '"blocking_plan_issues":[],"same_plan_followups":[],"future_followups":[]}\n'
-        "<!-- AGENT_PLAN_STATE: approved -->\n-- OpenAI Codex"
-    )
-
-    with pytest.raises(AgentLoopError, match="plan_review is missing required field"):
-        parse_plan_review(review, reviewer="OpenAI Codex")
-
-
-def test_extract_structured_plan_revision_payload_accepts_human_requirements_prefix():
-    payload = json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "plan_revision",
-            "state": "blocking",
-            "summary": "Revised the plan.",
-            "prior_plan_item_dispositions": [],
-            "plan_steps": ["Update protocol.py."],
-        }
-    )
-    text = (
-        payload
-        + "\n<!-- HUMAN_REQUIREMENTS_ADDRESSED -->\n\n### Human requirements\n- Requirement 1: covered in step 1.\n"
-        + "\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex"
-    )
-
-    assert _extract_structured_plan_revision_payload(text) is not None
-
-
-def test_extract_structured_plan_revision_payload_rejects_bad_footer_ordering():
-    payload = json.dumps(
-        {
-            "schema_version": 1,
-            "kind": "plan_revision",
-            "state": "blocking",
-            "summary": "Revised the plan.",
-            "prior_plan_item_dispositions": [],
-            "plan_steps": ["Update protocol.py."],
-        }
-    )
-    text = payload + "\n-- OpenAI Codex\n<!-- AGENT_PLAN_STATE: blocking -->"
-
-    with pytest.raises(AgentLoopError, match="AGENT_PLAN_STATE"):
-        _extract_structured_plan_revision_payload(text)
-
-
-def test_expect_string_list_enforces_min_length():
-    with pytest.raises(AgentLoopError, match="must contain at least 1 item"):
-        _expect_string_list([], context="plan_revision.plan_steps", item_context="plan_revision.plan_steps", min_length=1)
 
 
 def test_review_prompt_includes_prior_unresolved_items_and_disposition_instructions(tmp_path):
@@ -5177,392 +1312,6 @@ def test_structured_plan_review_preserves_human_requirements_resolution_marker()
     assert public.index("<!-- HUMAN_REQUIREMENTS_RESOLVED -->") < public.index("<!-- AGENT_PLAN_STATE: approved -->")
 
 
-def test_render_canonical_plan_steps_numbers_items():
-    assert render_canonical_plan_steps(("Update protocol.py.", "Add tests.")) == (
-        "1. Update protocol.py.\n2. Add tests."
-    )
-
-
-def test_render_canonical_plan_revision_and_public_comment():
-    parsed = validate_structured_plan_revision(
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_revision",
-                "state": "blocking",
-                "summary": "Revised the plan to cover rollback behavior.",
-                "prior_plan_item_dispositions": [
-                    {"item_id": "item-4", "disposition": "resolved", "note": "Added a resume-path step."}
-                ],
-                "plan_steps": ["Update protocol.py.", "Add orchestrator resume tests."],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex"
-    )
-    assert parsed is not None
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-4",
-            reviewer="OpenAI Codex",
-            source_round=2,
-            text="Add a resume-path step.",
-            status="blocking",
-        ),
-    )
-
-    canonical = render_canonical_plan_revision(parsed, prior_items)
-    public = _render_public_plan_revision_comment(
-        parsed,
-        prior_items=prior_items,
-        raw_text='{"schema_version":1}\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex',
-        agent="Codex",
-    )
-
-    assert canonical == (
-        "Revised the plan to cover rollback behavior.\n\n"
-        "### Prior plan item dispositions\n"
-        "- [item-4] Blocking issue from OpenAI Codex, round 2: Add a resume-path step. -> "
-        "resolved: Added a resume-path step.\n\n"
-        "### Plan steps\n"
-        "1. Update protocol.py.\n"
-        "2. Add orchestrator resume tests."
-    )
-    assert public == (
-        "## Revised plan\n\n"
-        + canonical
-        + "\n\n<!-- AGENT_PLAN_STATE: blocking -->\n\n-- OpenAI Codex"
-    )
-    assert '"kind": "plan_revision"' not in public
-
-
-def test_render_structured_plan_state_to_public_markdown():
-    raw = (
-        json.dumps(
-            {
-                "kind": "plan_state",
-                "summary": "Plan the renderer fix.",
-                "plan_steps": ["Detect structured plan_state.", "Render public markdown."],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: approved -->\n-- Google Antigravity"
-    )
-    parsed = validate_structured_plan_state(raw)
-
-    assert parsed is not None
-    public = render_public_agent_comment(
-        kind="plan_state",
-        parsed=parsed,
-        agent="antigravity",
-        model_used="Gemini 3.1 Pro (High)",
-    )
-
-    assert public == (
-        "## Plan\n\n"
-        "Plan the renderer fix.\n\n"
-        "### Plan steps\n"
-        "1. Detect structured plan_state.\n"
-        "2. Render public markdown.\n\n"
-        "<!-- AGENT_PLAN_STATE: approved -->\n\n"
-        "-- Google Antigravity: Gemini 3.1 Pro (High)"
-    )
-    assert '"kind": "plan_state"' not in public
-
-
-def test_render_public_coder_followup_comment():
-    parsed = validate_structured_coder_followup(
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Added the requested regression test.",
-                "addressed_items": ["item-1", "item-2"],
-                "remaining_items": [],
-                "addressed_item_notes": {
-                    "item-1": "Added coverage for the parser.",
-                    "item-2": "Updated the helper.",
-                },
-                "human_requirements": {
-                    "addressed_ids": ["Requirement 1"],
-                    "checked_discussion_directly": False,
-                },
-                "tests_run": [
-                    "python -m pytest tests/test_agent_loop.py -k coder_followup"
-                ],
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- Anthropic Claude"
-    )
-    assert parsed is not None
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-1",
-            reviewer="OpenAI Codex",
-            source_round=1,
-            text="Add a regression test before merge.",
-            status="blocking",
-        ),
-        UnresolvedReviewItem(
-            item_id="item-2",
-            reviewer="Google Gemini",
-            source_round=2,
-            text="Rename the shared helper.",
-            status="same-pr",
-        ),
-    )
-
-    rendered = _render_public_coder_followup_comment(
-        parsed,
-        agent="Claude",
-        prior_items=prior_items,
-    )
-
-    assert rendered == (
-        "## Coder follow-up\n\n"
-        "Added the requested regression test.\n\n"
-        "### Addressed items\n"
-        "- item-1: Blocking issue from OpenAI Codex, round 1: Add a regression test before merge.\n"
-        "  - Resolution: Added coverage for the parser.\n"
-        "- item-2: Same-PR follow-up from Google Gemini, round 2: Rename the shared helper.\n"
-        "  - Resolution: Updated the helper.\n\n"
-        "### Remaining items\n"
-        "- None.\n\n"
-        "### Tests run\n"
-        "- python -m pytest tests/test_agent_loop.py -k coder_followup\n\n"
-        "<!-- AGENT_STATE: blocking -->\n\n"
-        "-- Anthropic Claude"
-    )
-    assert "```json" not in rendered
-    assert '"kind": "coder_followup"' not in rendered
-
-    without_tests = validate_structured_coder_followup(
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Still working through the review.",
-                "addressed_items": [],
-                "remaining_items": ["item-3"],
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": False,
-                },
-                "tests_run": [],
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- Anthropic Claude"
-    )
-    assert without_tests is not None
-    rendered_without_tests = _render_public_coder_followup_comment(
-        without_tests,
-        agent="Claude",
-    )
-    assert "### Tests run" not in rendered_without_tests
-    assert "### Addressed items\n- None." in rendered_without_tests
-    assert (
-        "### Remaining items\n"
-        "- item-3: Item context unavailable in current round metadata.\n"
-        "  - Reason: No reason provided by coder."
-    ) in rendered_without_tests
-
-
-def test_render_public_coder_followup_comment_expands_carried_items_with_notes_and_placeholders():
-    parsed = validate_structured_coder_followup(
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Fixed the blocker and deferred the follow-up.",
-                "addressed_items": ["item-1"],
-                "remaining_items": ["item-2"],
-                "addressed_item_notes": {"item-1": "Restored the missing validation branch."},
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": False,
-                },
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- Anthropic Claude"
-    )
-    assert parsed is not None
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-1",
-            reviewer="OpenAI Codex",
-            source_round=2,
-            text="  - Preserve structured coder follow-up metadata.\n\nExtra context should be summarized.",
-            status="blocking",
-        ),
-        UnresolvedReviewItem(
-            item_id="item-2",
-            reviewer="Google Gemini",
-            source_round=3,
-            text="Move the rendering helper into a shared module.",
-            status="same-pr",
-        ),
-    )
-
-    rendered = _render_public_coder_followup_comment(
-        parsed,
-        agent="Claude",
-        prior_items=prior_items,
-    )
-
-    assert (
-        "- item-1: Blocking issue from OpenAI Codex, round 2: "
-        "Preserve structured coder follow-up metadata."
-    ) in rendered
-    assert "  - Resolution: Restored the missing validation branch." in rendered
-    assert (
-        "- item-2: Same-PR follow-up from Google Gemini, round 3: "
-        "Move the rendering helper into a shared module."
-    ) in rendered
-    assert "  - Reason: No reason provided by coder." in rendered
-
-
-def test_render_public_coder_followup_comment_expands_pr_220_remaining_items():
-    parsed = validate_structured_coder_followup(
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "coder_followup",
-                "state": "blocking",
-                "summary": "Hardened markdown stripping; two follow-ups remain.",
-                "addressed_items": ["item-3", "item-4"],
-                "remaining_items": ["item-5", "item-6"],
-                "remaining_item_notes": {
-                    "item-5": "Deferred because URL canonicalization needs product confirmation.",
-                    "item-6": "Deferred because the helper move should be isolated from this fix.",
-                },
-                "human_requirements": {
-                    "addressed_ids": [],
-                    "checked_discussion_directly": False,
-                },
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- Anthropic Claude"
-    )
-    assert parsed is not None
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-5",
-            reviewer="Google Gemini",
-            source_round=3,
-            text=(
-                "Update `server/static/index.html` and `server/static/landing.html` to use "
-                "relative paths for `og:image` and `og:url` if possible."
-            ),
-            status="same-pr",
-        ),
-        UnresolvedReviewItem(
-            item_id="item-6",
-            reviewer="Google Gemini",
-            source_round=3,
-            text=(
-                "Deduplicate `_strip_markdown` helper logic between `server/app.py` and "
-                "`core/orchestrator.py` by moving it to `core/utils.py`."
-            ),
-            status="same-pr",
-        ),
-    )
-
-    rendered = _render_public_coder_followup_comment(
-        parsed,
-        agent="Claude",
-        prior_items=prior_items,
-    )
-
-    assert "- item-5: Same-PR follow-up from Google Gemini, round 3:" in rendered
-    assert "relative paths" in rendered
-    assert "  - Reason: Deferred because URL canonicalization needs product confirmation." in rendered
-    assert "- item-6: Same-PR follow-up from Google Gemini, round 3:" in rendered
-    assert "Deduplicate `_strip_markdown` helper logic" in rendered
-    assert "  - Reason: Deferred because the helper move should be isolated from this fix." in rendered
-    assert "\n- item-5\n" not in rendered
-    assert "\n- item-6\n" not in rendered
-
-
-def test_render_public_plan_review_comment_normalizes_sections():
-    parsed = parse_structured_plan_review(
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "plan_review",
-                "state": "blocking",
-                "summary": "Still blocked on coverage.",
-                "blocking_plan_issues": ["Add a resume coverage test."],
-                "same_plan_followups": ["Mention canonical hashing explicitly."],
-                "future_followups": [],
-                "prior_plan_item_dispositions": [
-                    {"item_id": "item-2", "disposition": "same-plan", "note": "Still needs one more prompt assertion."}
-                ],
-            }
-        )
-        + "\n<!-- AGENT_PLAN_STATE: blocking -->\n-- OpenAI Codex",
-        reviewer="OpenAI Codex",
-    )
-    assert parsed is not None
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-2",
-            reviewer="Google Gemini",
-            source_round=1,
-            text="Mention canonical hashing explicitly.",
-            status="same-plan",
-        ),
-    )
-
-    rendered = _render_public_plan_review_comment(
-        parsed,
-        reviewer="OpenAI Codex",
-        prior_items=prior_items,
-        dispositions=parsed.dispositions,
-    )
-
-    assert rendered == (
-        "**Review verdict:** Blocking\n\n"
-        "Still blocked on coverage.\n\n"
-        "### Blocking plan issues\n"
-        "- Add a resume coverage test.\n\n"
-        "### Same-plan follow-ups\n"
-        "- Mention canonical hashing explicitly.\n\n"
-        "### Prior unresolved plan item dispositions\n"
-        "- [item-2] Same-plan follow-up from Google Gemini, round 1: Mention canonical hashing explicitly. -> "
-        "same-plan: Still needs one more prompt assertion.\n\n"
-        "<!-- AGENT_PLAN_STATE: blocking -->\n"
-        "-- OpenAI Codex"
-    )
-
-
-def test_review_freeform_summary_text_strips_structured_followup_sections():
-    review = """**Review verdict:** blocking
-
-Blocking issue summary.
-
-### Blocking issues
-- needs one more assertion
-
-### Prior unresolved item dispositions
-- [item-1] still blocking: needs one more assertion
-
-### Human requirements
-- Requirement 1: addressed in the latest patch
-
-### Same-PR follow-ups
-- Rename helper
-
-### Future follow-ups
-- Document cleanup later
-
-<!-- AGENT_STATE: blocking -->
--- OpenAI Codex
-"""
-
-    assert _review_freeform_summary_text(review) == "Blocking issue summary."
-
-
 def test_validate_review_response_accepts_structured_pr_review():
     review = (
         json.dumps(
@@ -5851,266 +1600,6 @@ def test_validate_coder_followup_response_requires_regular_synthetic_human_requi
             unresolved_items=unresolved_items,
             human_requirements=(),
         )
-
-
-def test_render_public_pr_review_comment_uses_normalized_sections_and_footer():
-    parsed = parse_review(
-        (
-            "Need one more regression test."
-            + blocking_issues("Exercise the structured-resume path.")
-            + "\n\n### Same-PR follow-ups\n- Rename the helper for clarity."
-            + prior_item_dispositions("[item-1] resolved")
-            + "\n<!-- HUMAN_REQUIREMENTS_RESOLVED -->\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-        ),
-        reviewer="OpenAI Codex",
-    )
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-1",
-            reviewer="Anthropic Claude",
-            source_round=1,
-            text="Add a regression test before merge.",
-            status="blocking",
-        ),
-    )
-
-    rendered = _render_public_pr_review_comment(
-        parsed,
-        reviewer="Codex",
-        human_requirements_resolved_flag=True,
-        prior_items=prior_items,
-        dispositions=parsed.dispositions,
-    )
-
-    assert rendered == (
-        "**Review verdict:** Blocking\n\n"
-        "Need one more regression test.\n\n"
-        "### Blocking issues\n"
-        "- Exercise the structured-resume path.\n\n"
-        "### Same-PR follow-ups\n"
-        "- Rename the helper for clarity.\n\n"
-        "### Prior unresolved item dispositions\n"
-        "- [item-1] Blocking issue from Anthropic Claude, round 1: Add a regression test before merge. -> resolved\n\n"
-        "<!-- HUMAN_REQUIREMENTS_RESOLVED -->\n"
-        "<!-- AGENT_STATE: blocking -->\n"
-        "-- OpenAI Codex"
-    )
-
-
-def test_render_public_pr_review_comment_normalizes_markdown_and_structured_reviews_the_same():
-    markdown_review = (
-        "Need one more regression test."
-        + blocking_issues("Exercise the structured-resume path.")
-        + "\n\n### Same-PR follow-ups\n- Rename the helper for clarity."
-        + prior_item_dispositions("[item-1] resolved")
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-    structured_review = (
-        json.dumps(
-            {
-                "schema_version": 1,
-                "kind": "pr_review",
-                "state": "blocking",
-                "summary": "Need one more regression test.",
-                "blocking_items": ["Exercise the structured-resume path."],
-                "same_pr_followups": ["Rename the helper for clarity."],
-                "future_followups": [],
-                "prior_item_dispositions": [{"item_id": "item-1", "disposition": "resolved"}],
-            }
-        )
-        + "\n<!-- AGENT_STATE: blocking -->\n-- OpenAI Codex"
-    )
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-1",
-            reviewer="Anthropic Claude",
-            source_round=1,
-            text="Add a regression test before merge.",
-            status="blocking",
-        ),
-    )
-
-    markdown_rendered = _render_public_pr_review_comment(
-        parse_review(markdown_review, reviewer="OpenAI Codex"),
-        reviewer="Codex",
-        human_requirements_resolved_flag=False,
-        prior_items=prior_items,
-        dispositions=parse_review(markdown_review, reviewer="OpenAI Codex").dispositions,
-    )
-    structured_parsed = parse_pr_review(structured_review, reviewer="OpenAI Codex")
-    structured_rendered = _render_public_pr_review_comment(
-        structured_parsed,
-        reviewer="Codex",
-        human_requirements_resolved_flag=False,
-        prior_items=prior_items,
-        dispositions=structured_parsed.dispositions,
-    )
-
-    assert markdown_rendered == structured_rendered
-
-
-def test_render_public_pr_review_comment_includes_visible_approved_verdict():
-    rendered = _render_public_pr_review_comment(
-        parse_review(
-            "Looks good to me.\n<!-- AGENT_STATE: approved -->\n-- OpenAI Codex",
-            reviewer="OpenAI Codex",
-        ),
-        reviewer="Codex",
-        human_requirements_resolved_flag=False,
-        prior_items=(),
-        dispositions=(),
-    )
-
-    assert rendered == (
-        "**Review verdict:** Approved\n\n"
-        "Looks good to me.\n\n"
-        "<!-- AGENT_STATE: approved -->\n"
-        "-- OpenAI Codex"
-    )
-
-
-def test_format_unresolved_item_label_normalizes_multiline_text_and_preserves_origin_status():
-    item = UnresolvedReviewItem(
-        item_id="item-7",
-        reviewer="Google Gemini",
-        source_round=1,
-        text="  - require source issue reference in PR body  \n\nUpdate from Anthropic Claude: keep the wording compact",
-        status="resolved",
-        source_status="same-pr",
-    )
-
-    assert _format_unresolved_item_label(item) == (
-        "Same-PR follow-up from Google Gemini, round 1: require source issue reference in PR body"
-    )
-
-
-def test_format_unresolved_item_label_truncates_at_fixed_limit():
-    summary = "a" * (ITEM_SUMMARY_LIMIT + 20)
-    item = UnresolvedReviewItem(
-        item_id="item-8",
-        reviewer="OpenAI Codex",
-        source_round=2,
-        text=summary,
-        status="blocking",
-    )
-
-    label = _format_unresolved_item_label(item)
-
-    assert label.startswith("Blocking issue from OpenAI Codex, round 2: ")
-    assert label.endswith("...")
-    rendered_summary = label.split(": ", 1)[1]
-    assert len(rendered_summary) == ITEM_SUMMARY_LIMIT
-
-
-def test_format_unresolved_item_label_special_cases_human_requirements_ack_item():
-    item = UnresolvedReviewItem(
-        item_id=HUMAN_REQUIREMENTS_ACK_ITEM_ID,
-        reviewer="Orchestrator",
-        source_round=3,
-        text="Coder response missing required `### Human requirements` section.",
-        status="blocking",
-    )
-
-    assert _format_unresolved_item_label(item) == (
-        "Human-requirements acknowledgement item, round 3: "
-        "Coder response missing required `### Human requirements` section."
-    )
-
-
-def test_render_public_review_comment_replaces_dispositions_without_exposing_same_round_new_items():
-    body = """Still blocked.
-
-### Same-PR follow-ups
-- Keep the source issue reference in the PR body.
-
-### Prior unresolved item dispositions
-- [item-1] same-pr
-
-<!-- AGENT_STATE: blocking -->
--- OpenAI Codex
-"""
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-1",
-            reviewer="Google Gemini",
-            source_round=1,
-            text="Require source issue reference in PR body.\n\nUpdate from Anthropic Claude: keep the note compact",
-            status="same-pr",
-        ),
-    )
-    dispositions = parse_unresolved_item_dispositions(
-        prior_item_dispositions("[item-1] Same-PR follow-up from Google Gemini, round 1: ignored by parser -> same-pr: keep the body reference"),
-        reviewer="OpenAI Codex",
-    )
-    new_items = (
-        UnresolvedReviewItem(
-            item_id="item-2",
-            reviewer="OpenAI Codex",
-            source_round=2,
-            text="Keep the source issue reference in the PR body.",
-            status="same-pr",
-        ),
-    )
-
-    rendered = _render_public_review_comment(
-        body,
-        review_kind="pr",
-        prior_items=prior_items,
-        dispositions=dispositions,
-        new_items=new_items,
-    )
-
-    assert "### Same-PR follow-ups\n- Keep the source issue reference in the PR body." in rendered
-    assert (
-        "### Prior unresolved item dispositions\n"
-        "- [item-1] Same-PR follow-up from Google Gemini, round 1: Require source issue reference in PR body. -> same-pr: keep the body reference"
-    ) in rendered
-    assert "### New tracked unresolved items" not in rendered
-    assert "[item-2]" not in rendered
-    assert rendered.rstrip().endswith("-- OpenAI Codex")
-
-
-def test_render_public_review_comment_preserves_unknown_disposition_values():
-    body = """Still blocked.
-
-### Prior unresolved item dispositions
-- [item-1] same-pr
-
-<!-- AGENT_STATE: blocking -->
--- OpenAI Codex
-"""
-    prior_items = (
-        UnresolvedReviewItem(
-            item_id="item-1",
-            reviewer="Google Gemini",
-            source_round=1,
-            text="Keep the parser and renderer aligned when new dispositions are added.",
-            status="same-pr",
-        ),
-    )
-    dispositions = (
-        ReviewItemDisposition(
-            item_id="item-1",
-            reviewer="OpenAI Codex",
-            disposition="deferred",
-            note="tracked for a later parser update",
-        ),
-    )
-
-    rendered = _render_public_review_comment(
-        body,
-        review_kind="pr",
-        prior_items=prior_items,
-        dispositions=dispositions,
-        new_items=(),
-    )
-
-    assert (
-        "### Prior unresolved item dispositions\n"
-        "- [item-1] Same-PR follow-up from Google Gemini, round 1: "
-        "Keep the parser and renderer aligned when new dispositions are added. "
-        "-> deferred: tracked for a later parser update"
-    ) in rendered
 
 
 def test_apply_unresolved_item_dispositions_appends_disposition_notes_to_text():
